ADD gemma dense omni mdality model
