+++ src/megatron/bridge/models/gemma_vl/__init__.py
@@ -15,7 +15,7 @@
 from megatron.bridge.models.gemma_vl.gemma3_vl_bridge import Gemma3VLBridge
 from megatron.bridge.models.gemma_vl.gemma3_vl_provider import Gemma3VLModelProvider
 from megatron.bridge.models.gemma_vl.gemma4_vl_bridge import Gemma4VLBridge
-from megatron.bridge.models.gemma_vl.gemma4_vl_provider import Gemma4E4BVLProvider, Gemma4VLModelProvider
+from megatron.bridge.models.gemma_vl.gemma4_vl_provider import Gemma4DenseVLProvider, Gemma4VLModelProvider
 from megatron.bridge.models.gemma_vl.modeling_gemma3_vl import Gemma3VLModel
 from megatron.bridge.models.gemma_vl.modeling_gemma4_vl import Gemma4VLModel
 
@@ -27,5 +27,5 @@ __all__ = [
     "Gemma4VLModel",
     "Gemma4VLBridge",
     "Gemma4VLModelProvider",
-    "Gemma4E4BVLProvider",
+    "Gemma4DenseVLProvider",
 ]
