+++ src/megatron/bridge/models/gemma_vl/gemma4_vl_bridge.py
@@ -13,25 +13,29 @@
 # limitations under the License.
 
 """
-Megatron Bridge for Gemma 4 VL (Vision-Language).
+Megatron Bridge for Gemma 4 (CausalLM text-only and ConditionalGeneration VL).
 
-Extends the Gemma 4 text bridge to handle the full VLM checkpoint with
-vision tower, multimodal embedder, and language model.
+Supports all Gemma 4 variants:
+  - MoE (``enable_moe_block=True``): ``Gemma4ForCausalLM`` / ``Gemma4ForConditionalGeneration``
+  - Dense (``enable_moe_block=False``): same HF classes, dispatched via ``Gemma4DenseProvider``
 
-Weight prefixes in HF VLM checkpoint (after stripping outer ``model.``):
-- ``language_model.layers.*``     → language model decoder layers
-- ``language_model.embed_tokens`` → language model embedding
-- ``language_model.norm``         → final layernorm
-- ``vision_tower.*``              → HF vision encoder (replicated)
-- ``embed_vision.*``              → multimodal projector (replicated)
+Bridge conversion architecture:
+  AutoBridge.from_hf_pretrained("google/gemma-4-E4B-it")
+    └─ Gemma4VLBridge (registered for Gemma4ForConditionalGeneration)
+         ├─ provider_bridge()  text mode  → Gemma4DenseProvider (pretraining)
+         │                     auto/vl    → Gemma4DenseVLProvider (full VL)
+         └─ _dense_mapping_registry() / _moe_vl_mapping_registry()
+
+  AutoBridge.from_hf_pretrained("google/gemma-4-26B-A4B")
+    └─ Gemma4Bridge (registered for Gemma4ForCausalLM, MoE or Dense)
 """
 
-from dataclasses import fields
 import os
 import re
-from typing import Mapping
+from typing import Any, Mapping
 
 import torch
+from megatron.core.models.gpt.gpt_model import GPTModel
 
 from megatron.bridge.models.conversion.mapping_registry import MegatronMappingRegistry
 from megatron.bridge.models.conversion.model_bridge import MegatronModelBridge
@@ -40,6 +44,7 @@ from megatron.bridge.models.conversion.param_mapping import (
     FusedExpertMapping,
     FusedGatedExpertMapping,
     GatedMLPMapping,
+    QKVMapping,
     ReplicatedMapping,
     split_qkv_weights,
 )
@@ -48,174 +53,183 @@ from megatron.bridge.models.conversion.transformers_compat import (
     rope_local_base_freq_from_hf,
     rope_theta_from_hf,
 )
-from megatron.bridge.models.gemma.gemma4_bridge import (
-    Gemma4Bridge,
-    _Gemma4QKVMapping,
-    _infer_attn_pattern,
+from megatron.bridge.models.gemma_vl.gemma4_vl_provider import (
+    Gemma4DenseVLProvider,
+    Gemma4ModelProvider,
+    Gemma4VLModelProvider,
 )
-from megatron.bridge.models.gemma.gemma4_layer_specs import Gemma4E4BProvider
-from megatron.bridge.models.gemma_vl.gemma4_vl_provider import Gemma4E4BVLProvider, Gemma4VLModelProvider
-from megatron.bridge.models.gemma_vl.modeling_gemma4_vl import Gemma4VLModel
+from megatron.bridge.models.gemma_vl.modeling_gemma4_vl import Gemma4DenseProvider, Gemma4VLModel
+from megatron.bridge.models.hf_pretrained.causal_lm import PreTrainedCausalLM
 from megatron.bridge.models.hf_pretrained.vlm import PreTrainedVLM
 
 
+# Register Gemma4 custom module types for AutoMapping
+AutoMapping.register_module_type("Gemma4TEDotProductAttention", "replicated")
+AutoMapping.register_module_type("Gemma4SelfAttention", "replicated")
+AutoMapping.register_module_type("Gemma4TransformerLayer", "replicated")
+AutoMapping.register_module_type("Gemma4TopKRouter", "replicated")
+AutoMapping.register_module_type("Gemma4MoELayer", "replicated")
+AutoMapping.register_module_type("SharedExpertMLP", "column")
+
+
+class _Gemma4QKVMapping(QKVMapping):
+    """QKV mapping tolerating missing v_proj on global attention layers (K=V)."""
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.allow_hf_name_mismatch = True
+
+
+class _Gemma4DenseQKVMapping(QKVMapping):
+    """QKV mapping tolerating missing k_proj AND v_proj on shared-KV layers."""
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.allow_hf_name_mismatch = True
+
+
+def _infer_attn_pattern(layer_types: list[str]) -> tuple[int, int]:
+    """Infer (sliding, global) interleaved attention pattern from layer_types list."""
+    for i, lt in enumerate(layer_types):
+        if lt == "full_attention":
+            sliding_count = i
+            full_count = 0
+            for j in range(i, len(layer_types)):
+                if layer_types[j] == "full_attention":
+                    full_count += 1
+                else:
+                    break
+            return (sliding_count, full_count)
+    return (len(layer_types), 0)
+
+
+# ---------------------------------------------------------------------------
+# Gemma4Bridge — text-only CausalLM bridge (MoE and Dense)
+# ---------------------------------------------------------------------------
+
+
 @MegatronModelBridge.register_bridge(
-    source="Gemma4ForConditionalGeneration",
-    target=Gemma4VLModel,
-    provider=Gemma4VLModelProvider,
-    model_type="gemma4_vl",
+    source="Gemma4ForCausalLM",
+    target=GPTModel,
+    provider=Gemma4ModelProvider,
+    model_type="gemma4",
 )
-class Gemma4VLBridge(MegatronModelBridge):
-    """Megatron Bridge for Gemma 4 Vision-Language models.
+class Gemma4Bridge(MegatronModelBridge):
+    """Megatron Bridge for Gemma 4 text-only (CausalLM).
 
-    Handles conversion between HuggingFace Gemma4ForConditionalGeneration and
-    Megatron-Core Gemma4VLModel.
-
-    Example:
-        >>> from megatron.bridge import AutoBridge
-        >>> bridge = AutoBridge.from_hf_pretrained("google/gemma-4-26B-A4B")
-        >>> provider = bridge.to_megatron_provider()
+    Dispatches to Dense or MoE path based on ``enable_moe_block`` in HF config.
     """
 
-    def provider_bridge(self, hf_pretrained: PreTrainedVLM) -> Gemma4VLModelProvider | Gemma4E4BVLProvider | Gemma4E4BProvider:
+    _CONDITIONAL_MOE_FIELDS = frozenset({"num_moe_experts", "moe_router_topk", "moe_ffn_hidden_size"})
+
+    def _should_map_hf_config_field(self, hf_config: Any, hf_name: str, megatron_name: str, value: Any) -> bool:
+        if megatron_name in self._CONDITIONAL_MOE_FIELDS:
+            return getattr(hf_config, "enable_moe_block", True)
+        return super()._should_map_hf_config_field(hf_config, hf_name, megatron_name, value)
+
+    def provider_bridge(self, hf_pretrained: PreTrainedCausalLM) -> "Gemma4ModelProvider | Gemma4DenseProvider":
         hf_config = hf_pretrained.config
-        text_config = hf_config.text_config
-        vision_config = hf_config.vision_config
+        if not getattr(hf_config, "enable_moe_block", False):
+            self._is_dense = True
+            return self._build_dense_provider(hf_config)
 
-        if not getattr(text_config, "enable_moe_block", False):
-            # Dense E4B path: use full VL by default, but allow text-only
-            # conversion for text pretraining from a ConditionalGeneration HF config.
-            self._is_dense_e4b = True
-            self._is_dense_e4b_text_only = self._conversion_mode() == "text"
-            if self._is_dense_e4b_text_only:
-                return Gemma4Bridge._build_dense_e4b_provider(self, text_config)
-            return self._build_dense_e4b_vl_provider(hf_config, text_config, vision_config)
-        self._is_dense_e4b = False
-        self._is_dense_e4b_text_only = False
+        self._is_dense = False
+        return self._build_moe_provider(hf_config)
 
-        # Use base class helper for common config conversion from text_config
-        provider_kwargs = self.hf_config_to_provider_kwargs(text_config)
-        provider = Gemma4VLModelProvider(**provider_kwargs)
+    def _build_dense_provider(self, hf_config) -> Gemma4DenseProvider:
+        """Build a Gemma4DenseProvider from HF config."""
+        rope_params = getattr(hf_config, "rope_parameters", {}) or {}
+        sliding_rope = rope_params.get("sliding_attention", {})
+        full_rope = rope_params.get("full_attention", {})
 
-        # === Gemma 4 text-specific features (same as Gemma4Bridge) ===
-        provider.window_size = getattr(text_config, "sliding_window", 1024)
+        layer_types = getattr(hf_config, "layer_types", None)
+        if layer_types is not None:
+            layer_types = [layer_type == "sliding_attention" for layer_type in layer_types]
 
-        # Dual RoPE bases
-        provider.rotary_base = (
-            rope_local_base_freq_from_hf(text_config),
-            rope_theta_from_hf(text_config),
+        return Gemma4DenseProvider(
+            num_layers=hf_config.num_hidden_layers,
+            hidden_size=hf_config.hidden_size,
+            ffn_hidden_size=hf_config.intermediate_size,
+            num_attention_heads=hf_config.num_attention_heads,
+            num_query_groups=hf_config.num_key_value_heads,
+            kv_channels=getattr(hf_config, "head_dim", 256),
+            global_kv_channels=getattr(hf_config, "global_head_dim", 512),
+            num_global_query_groups=getattr(
+                hf_config,
+                "num_global_key_value_heads",
+                getattr(hf_config, "num_key_value_heads", 2),
+            ),
+            seq_length=hf_config.max_position_embeddings,
+            vocab_size=hf_config.vocab_size,
+            normalization="RMSNorm",
+            layernorm_epsilon=hf_config.rms_norm_eps,
+            window_attn_skip_freq=layer_types if layer_types is not None else 6,
+            sliding_window_rope_base=sliding_rope.get("rope_theta", 10000.0),
+            full_attention_rope_base=full_rope.get("rope_theta", 1000000.0),
+            full_attention_rope_partial_factor=full_rope.get("partial_rotary_factor", 0.25),
+            num_kv_shared_layers=getattr(hf_config, "num_kv_shared_layers", 0),
+            per_layer_embed_vocab_size=getattr(
+                hf_config, "vocab_size_per_layer_input", hf_config.vocab_size
+            ),
+            per_layer_embed_dim=getattr(hf_config, "hidden_size_per_layer_input", 256),
+            bf16=True,
         )
 
-        # QK norm
-        head_dim = getattr(text_config, "head_dim", 256)
+    def _build_moe_provider(self, hf_config) -> Gemma4ModelProvider:
+        """Build a Gemma4ModelProvider from HF config (MoE path)."""
+        provider_kwargs = self.hf_config_to_provider_kwargs(hf_config)
+        provider = Gemma4ModelProvider(**provider_kwargs)
+
+        provider.window_size = getattr(hf_config, "sliding_window", 1024)
+        provider.rotary_base = (
+            rope_local_base_freq_from_hf(hf_config),
+            rope_theta_from_hf(hf_config),
+        )
+
+        head_dim = getattr(hf_config, "head_dim", 256)
         provider.softmax_scale = 1.0
         provider.kv_channels = head_dim
         provider.qk_layernorm = True
 
-        # Global attention overrides
-        provider.global_head_dim = getattr(text_config, "global_head_dim", 512)
-        provider.num_global_key_value_heads = getattr(text_config, "num_global_key_value_heads", 2)
+        provider.global_head_dim = getattr(hf_config, "global_head_dim", 512)
+        provider.num_global_key_value_heads = getattr(hf_config, "num_global_key_value_heads", 2)
 
-        provider.attention_k_eq_v = getattr(text_config, "attention_k_eq_v", False)
-
-        # Parse partial_rotary_factor
-        rope_params = getattr(text_config, "rope_parameters", {})
+        rope_params = getattr(hf_config, "rope_parameters", {})
         if isinstance(rope_params, dict):
             full_attn_rope = rope_params.get("full_attention", {})
             provider.global_rotary_percent = full_attn_rope.get("partial_rotary_factor", 0.25)
 
-        # Sliding/global layer pattern
-        layer_types = getattr(text_config, "layer_types", None)
+        layer_types = getattr(hf_config, "layer_types", None)
         if layer_types:
             provider.interleaved_attn_pattern = _infer_attn_pattern(layer_types)
 
-        # MoE MLP configuration
-        is_moe = getattr(text_config, "enable_moe_block", False)
-        if is_moe:
-            provider.num_moe_experts = getattr(text_config, "num_experts", None) or 128
-            provider.moe_router_topk = getattr(text_config, "top_k_experts", None) or 8
-            provider.moe_ffn_hidden_size = getattr(text_config, "moe_intermediate_size", None) or 704
-            provider.moe_shared_expert_intermediate_size = getattr(text_config, "intermediate_size", 2112)
+        if getattr(hf_config, "enable_moe_block", False):
+            provider.num_moe_experts = getattr(hf_config, "num_experts", 128)
+            provider.moe_router_topk = getattr(hf_config, "top_k_experts", 8)
+            provider.moe_ffn_hidden_size = getattr(hf_config, "moe_intermediate_size", 704)
+            provider.moe_shared_expert_intermediate_size = getattr(hf_config, "intermediate_size", 2112)
             provider.moe_shared_expert_overlap = False
             provider.moe_shared_expert_gate = False
             provider.moe_layer_freq = 1
 
-        # Logit softcapping
-        provider.final_logit_softcapping = getattr(text_config, "final_logit_softcapping", 30.0)
-
-        # Override dtype and vocab settings
+        provider.final_logit_softcapping = getattr(hf_config, "final_logit_softcapping", 30.0)
         provider.bf16 = True
         provider.params_dtype = torch.bfloat16
         provider.autocast_dtype = torch.bfloat16
         provider.make_vocab_size_divisible_by = 128
 
-        # === VL-specific config ===
-        provider.vision_config = vision_config
-        provider.text_config = text_config
-        provider.vision_soft_tokens_per_image = getattr(hf_config, "vision_soft_tokens_per_image", 280)
-
-        # Token IDs
-        provider.bos_token_id = getattr(hf_config, "bos_token_id", 2)
-        provider.eos_token_id = getattr(hf_config, "eos_token_id", 1)
-        provider.image_token_id = getattr(hf_config, "image_token_id", 258_880)
-        provider.video_token_id = getattr(hf_config, "video_token_id", 258_884)
-
         return provider
 
-    def _conversion_mode(self) -> str:
-        mode = getattr(self, "gemma4_conversion_mode", None) or os.environ.get("GEMMA4_CONVERSION_MODE", "auto")
-        mode = mode.lower()
-        if mode not in {"auto", "text", "vl"}:
-            raise ValueError(f"Invalid GEMMA4_CONVERSION_MODE={mode!r}; expected auto, text, or vl.")
-        return mode
-
-    def _build_dense_e4b_vl_provider(self, hf_config, text_config, vision_config) -> Gemma4E4BVLProvider:
-        """Build a Dense E4B VL provider while reusing the text Dense provider setup."""
-        text_provider = Gemma4Bridge._build_dense_e4b_provider(self, text_config)
-        provider = Gemma4E4BVLProvider()
-        for field in fields(Gemma4E4BProvider):
-            setattr(provider, field.name, getattr(text_provider, field.name))
-
-        provider.vision_config = vision_config
-        provider.text_config = text_config
-        provider.vision_soft_tokens_per_image = getattr(hf_config, "vision_soft_tokens_per_image", 280)
-        provider.bos_token_id = getattr(hf_config, "bos_token_id", 2)
-        provider.eos_token_id = getattr(hf_config, "eos_token_id", 1)
-        provider.image_token_id = getattr(hf_config, "image_token_id", 258_880)
-        provider.video_token_id = getattr(hf_config, "video_token_id", 258_884)
-        return provider
-
-    def maybe_modify_converted_hf_weight(
-        self,
-        task,
-        converted_weights_dict,
-        hf_state_dict,
-    ):
-        """Un-fuse fused weights and drop synthesized keys on export.
-
-        On import, ``maybe_modify_loaded_hf_weight`` applies two non-trivial fusions
-        to the MoE layers to simplify the MCore forward pass:
-
-        1. **Router fusion**: ``mg = hf * (scale * sqrt_hidden⁻¹ / pffl2)``
-        2. **Shared-expert gate/up fusion**: ``mg = hf * (pffl / pffl2)``
-
-        On export (Megatron → HF), this method inverts both fusions so the
-        resulting HF weights exactly match the original checkpoint.  It also
-        drops the synthesized ``v_proj`` key produced by ``QKVMapping.megatron_to_hf``
-        for K=V global-attention layers where ``v_proj`` is absent in HF.
-        """
+    def maybe_modify_converted_hf_weight(self, task, converted_weights_dict, hf_state_dict):
+        """Un-fuse fused weights and drop synthesized keys on export."""
         if not hf_state_dict:
             return converted_weights_dict
 
         result = {}
         for hf_name, tensor in converted_weights_dict.items():
-            # Drop synthesized v_proj (absent for K=V global-attention layers)
             if hf_name not in hf_state_dict:
                 continue
 
-            # ── Router weight inverse: mg = hf * (scale * hidden^-0.5 / pffl2)
-            #                         hf = mg / (scale * hidden^-0.5 / pffl2)
-            #                            = mg * pffl2 / (scale * hidden^-0.5)
             if hf_name.endswith("router.proj.weight"):
                 layer_match = re.search(r"layers\.(\d+)\.", hf_name)
                 if layer_match:
@@ -228,11 +242,9 @@ class Gemma4VLBridge(MegatronModelBridge):
                         ln2_weight = hf_state_dict[ln2_key].float().to(tensor.device)
                         hidden_size = tensor.shape[-1]
                         scalar_root_size = hidden_size**-0.5
-                        fusion_factor = router_scale * scalar_root_size / ln2_weight  # [hidden]
+                        fusion_factor = router_scale * scalar_root_size / ln2_weight
                         tensor = (tensor.float() / fusion_factor.unsqueeze(0)).to(tensor.dtype)
 
-            # ── Shared-expert gate/up inverse: mg = hf * (pffl / pffl2)
-            #                                  hf = mg * (pffl2 / pffl)
             elif hf_name.endswith(("mlp.gate_proj.weight", "mlp.up_proj.weight")) and "experts" not in hf_name:
                 layer_match = re.search(r"layers\.(\d+)\.", hf_name)
                 if layer_match:
@@ -243,7 +255,7 @@ class Gemma4VLBridge(MegatronModelBridge):
                     if pffl_key in hf_state_dict and pffl2_key in hf_state_dict:
                         w_pffl = hf_state_dict[pffl_key].float().to(tensor.device)
                         w_pffl2 = hf_state_dict[pffl2_key].float().to(tensor.device)
-                        correction = w_pffl / w_pffl2  # [hidden]
+                        correction = w_pffl / w_pffl2
                         tensor = (tensor.float() / correction.unsqueeze(0)).to(tensor.dtype)
 
             result[hf_name] = tensor
@@ -253,17 +265,424 @@ class Gemma4VLBridge(MegatronModelBridge):
     def maybe_modify_loaded_hf_weight(
         self, hf_param: str | dict[str, str], hf_state_dict: Mapping[str, torch.Tensor]
     ) -> torch.Tensor:
-        """Handle special weight loading for Gemma 4 VLM.
+        """Handle special weight loading for Gemma 4."""
+        if isinstance(hf_param, dict) and "v" in hf_param:
+            k_name = hf_param["k"]
+            v_name = hf_param["v"]
+            q_name = hf_param["q"]
 
-        K=V synthesis for global attention layers, router weight fusion, and
-        shared expert pre-norm fusion.
+            if k_name not in hf_state_dict and v_name not in hf_state_dict:
+                q_weight = hf_state_dict[q_name]
+                num_q_heads = 8
+                kv_head_dim = q_weight.shape[0] // num_q_heads
+                num_kv_heads = 2
+                kv_shape = (num_kv_heads * kv_head_dim, q_weight.shape[1])
+                k_zero = torch.zeros(kv_shape, dtype=q_weight.dtype, device=q_weight.device)
+                return {"q": q_weight, "k": k_zero, "v": torch.zeros_like(k_zero)}
 
-        HF param names have ``model.language_model.`` prefix (raw safetensors
-        keys include the outer ``model.`` from Gemma4ForConditionalGeneration).
-        """
-        # Dense E4B shared-KV layers omit both k_proj and v_proj in HF.  The
-        # Megatron model wires these layers to their source KV layers at runtime,
-        # so zero K/V rows are valid placeholders during checkpoint import.
+            if v_name not in hf_state_dict and k_name in hf_state_dict:
+                hf_weights = {}
+                for role, name in hf_param.items():
+                    if role == "v":
+                        hf_weights[role] = hf_state_dict[k_name].clone()
+                    else:
+                        hf_weights[role] = hf_state_dict[name]
+                return hf_weights
+
+        if isinstance(hf_param, dict) and "gate" in hf_param:
+            gate_name = hf_param["gate"]
+            if "mlp.gate_proj" in gate_name:
+                return self._fuse_shared_expert_prenorm(hf_param, hf_state_dict)
+
+        if isinstance(hf_param, str) and hf_param.endswith("router.proj.weight"):
+            return self._fuse_router_weight(hf_param, hf_state_dict)
+
+        return super().maybe_modify_loaded_hf_weight(hf_param, hf_state_dict)
+
+    def _fuse_router_weight(self, hf_param: str, hf_state_dict: Mapping[str, torch.Tensor]) -> torch.Tensor:
+        proj_weight = hf_state_dict[hf_param]
+        layer_match = re.search(r"layers\.(\d+)\.", hf_param)
+        if layer_match is None:
+            return proj_weight
+        layer_idx = layer_match.group(1)
+        scale_key = f"model.layers.{layer_idx}.router.scale"
+        ln2_key = f"model.layers.{layer_idx}.pre_feedforward_layernorm_2.weight"
+        if scale_key not in hf_state_dict or ln2_key not in hf_state_dict:
+            return proj_weight
+        router_scale = hf_state_dict[scale_key].float()
+        ln2_weight = hf_state_dict[ln2_key].float()
+        hidden_size = proj_weight.shape[-1]
+        scalar_root_size = hidden_size**-0.5
+        fusion_factor = router_scale * scalar_root_size / ln2_weight
+        fused_weight = proj_weight.float() * fusion_factor.unsqueeze(0)
+        return fused_weight.to(proj_weight.dtype)
+
+    def _fuse_shared_expert_prenorm(
+        self, hf_param: dict[str, str], hf_state_dict: Mapping[str, torch.Tensor]
+    ) -> dict[str, torch.Tensor]:
+        gate_name = hf_param["gate"]
+        layer_match = re.search(r"layers\.(\d+)\.", gate_name)
+        if layer_match is None:
+            return {role: hf_state_dict[name] for role, name in hf_param.items()}
+        layer_idx = layer_match.group(1)
+        pffl_key = f"model.layers.{layer_idx}.pre_feedforward_layernorm.weight"
+        pffl2_key = f"model.layers.{layer_idx}.pre_feedforward_layernorm_2.weight"
+        if pffl_key not in hf_state_dict or pffl2_key not in hf_state_dict:
+            return {role: hf_state_dict[name] for role, name in hf_param.items()}
+        w_pffl = hf_state_dict[pffl_key].float()
+        w_pffl2 = hf_state_dict[pffl2_key].float()
+        correction = w_pffl / w_pffl2
+        hf_weights = {}
+        for role, name in hf_param.items():
+            weight = hf_state_dict[name]
+            fused = weight.float() * correction.unsqueeze(0)
+            hf_weights[role] = fused.to(weight.dtype)
+        return hf_weights
+
+    def mapping_registry(self) -> MegatronMappingRegistry:
+        if getattr(self, "_is_dense", False):
+            return self._dense_mapping_registry()
+        return self._moe_mapping_registry()
+
+    def _dense_mapping_registry(self, megatron_prefix: str = "") -> MegatronMappingRegistry:
+        """Parameter mappings for the Dense variant."""
+        mp = megatron_prefix
+        hp = self._hf_layer_prefix()
+        param_mappings = {
+            f"{mp}embedding.word_embeddings.weight": f"{hp}embed_tokens.weight",
+            f"{mp}decoder.final_layernorm.weight": f"{hp}norm.weight",
+            f"{mp}per_layer_embedding.weight": f"{hp}embed_tokens_per_layer.weight",
+            f"{mp}per_layer_model_proj.weight": f"{hp}per_layer_model_projection.weight",
+            f"{mp}decoder.layers.*.input_layernorm.weight": f"{hp}layers.*.input_layernorm.weight",
+            f"{mp}decoder.layers.*.post_self_attn_layernorm.weight": f"{hp}layers.*.post_attention_layernorm.weight",
+            f"{mp}decoder.layers.*.pre_mlp_layernorm.weight": f"{hp}layers.*.pre_feedforward_layernorm.weight",
+            f"{mp}decoder.layers.*.post_mlp_layernorm.weight": f"{hp}layers.*.post_feedforward_layernorm.weight",
+            f"{mp}decoder.layers.*.self_attention.q_layernorm.weight": f"{hp}layers.*.self_attn.q_norm.weight",
+            f"{mp}decoder.layers.*.self_attention.k_layernorm.weight": f"{hp}layers.*.self_attn.k_norm.weight",
+            f"{mp}decoder.layers.*.self_attention.linear_proj.weight": f"{hp}layers.*.self_attn.o_proj.weight",
+            f"{mp}decoder.layers.*.mlp.linear_fc2.weight": f"{hp}layers.*.mlp.down_proj.weight",
+        }
+        mapping_list = [AutoMapping(megatron_param=m, hf_param=h) for m, h in param_mappings.items()]
+
+        mapping_list.append(
+            ReplicatedMapping(
+                megatron_param=f"{mp}per_layer_proj_norm.weight",
+                hf_param=f"{hp}per_layer_projection_norm.weight",
+            )
+        )
+        mapping_list.extend([
+            ReplicatedMapping(
+                megatron_param=f"{mp}decoder.layers.*.per_layer_input_gate.weight",
+                hf_param=f"{hp}layers.*.per_layer_input_gate.weight",
+            ),
+            ReplicatedMapping(
+                megatron_param=f"{mp}decoder.layers.*.per_layer_projection.weight",
+                hf_param=f"{hp}layers.*.per_layer_projection.weight",
+            ),
+            ReplicatedMapping(
+                megatron_param=f"{mp}decoder.layers.*.post_per_layer_input_norm.weight",
+                hf_param=f"{hp}layers.*.post_per_layer_input_norm.weight",
+            ),
+            ReplicatedMapping(
+                megatron_param=f"{mp}decoder.layers.*.layer_scalar",
+                hf_param=f"{hp}layers.*.layer_scalar",
+            ),
+            _Gemma4DenseQKVMapping(
+                megatron_param=f"{mp}decoder.layers.*.self_attention.linear_qkv.weight",
+                q=f"{hp}layers.*.self_attn.q_proj.weight",
+                k=f"{hp}layers.*.self_attn.k_proj.weight",
+                v=f"{hp}layers.*.self_attn.v_proj.weight",
+            ),
+            GatedMLPMapping(
+                megatron_param=f"{mp}decoder.layers.*.mlp.linear_fc1.weight",
+                gate=f"{hp}layers.*.mlp.gate_proj.weight",
+                up=f"{hp}layers.*.mlp.up_proj.weight",
+            ),
+        ])
+        return MegatronMappingRegistry(*mapping_list)
+
+    def _hf_layer_prefix(self) -> str:
+        """Text-only CausalLM: weights at ``model.*``; override in VL subclass."""
+        return "model."
+
+    def _moe_mapping_registry(self) -> MegatronMappingRegistry:
+        """Parameter mappings for the MoE variant."""
+        param_mappings = {
+            "embedding.word_embeddings.weight": "model.embed_tokens.weight",
+            "decoder.final_layernorm.weight": "model.norm.weight",
+            "decoder.layers.*.self_attention.linear_qkv.layer_norm_weight": "model.layers.*.input_layernorm.weight",
+            "decoder.layers.*.input_layernorm.weight": "model.layers.*.input_layernorm.weight",
+            "decoder.layers.*.self_attention.q_layernorm.weight": "model.layers.*.self_attn.q_norm.weight",
+            "decoder.layers.*.self_attention.k_layernorm.weight": "model.layers.*.self_attn.k_norm.weight",
+            "decoder.layers.*.self_attention.linear_proj.weight": "model.layers.*.self_attn.o_proj.weight",
+            "decoder.layers.*.self_attention.linear_proj.post_layernorm.weight": (
+                "model.layers.*.post_attention_layernorm.weight"
+            ),
+            "decoder.layers.*.pre_mlp_layernorm.weight": "model.layers.*.pre_feedforward_layernorm_2.weight",
+            "decoder.layers.*.mlp.shared_experts.linear_fc2.weight": "model.layers.*.mlp.down_proj.weight",
+            "decoder.layers.*.mlp.shared_experts.linear_fc2.post_layernorm.weight": (
+                "model.layers.*.post_feedforward_layernorm_1.weight"
+            ),
+            "decoder.layers.*.mlp.router.weight": "model.layers.*.router.proj.weight",
+            "decoder.layers.*.mlp.linear_fc2.weight": "model.layers.*.mlp.down_proj.weight",
+            "decoder.layers.*.mlp.linear_fc1.layer_norm_weight": "model.layers.*.post_attention_layernorm.weight",
+        }
+
+        mapping_list = [AutoMapping(megatron_param=m, hf_param=h) for m, h in param_mappings.items()]
+        mapping_list.extend([
+            _Gemma4QKVMapping(
+                megatron_param="decoder.layers.*.self_attention.linear_qkv.weight",
+                q="model.layers.*.self_attn.q_proj.weight",
+                k="model.layers.*.self_attn.k_proj.weight",
+                v="model.layers.*.self_attn.v_proj.weight",
+            ),
+            GatedMLPMapping(
+                megatron_param="decoder.layers.*.mlp.shared_experts.linear_fc1.weight",
+                gate="model.layers.*.mlp.gate_proj.weight",
+                up="model.layers.*.mlp.up_proj.weight",
+            ),
+            GatedMLPMapping(
+                megatron_param="decoder.layers.*.mlp.linear_fc1.weight",
+                gate="model.layers.*.mlp.gate_proj.weight",
+                up="model.layers.*.mlp.up_proj.weight",
+            ),
+            FusedGatedExpertMapping(
+                megatron_param="decoder.layers.*.mlp.experts.linear_fc1.weight*",
+                hf_param="model.layers.*.experts.gate_up_proj",
+            ),
+            FusedExpertMapping(
+                megatron_param="decoder.layers.*.mlp.experts.linear_fc2.weight*",
+                hf_param="model.layers.*.experts.down_proj",
+            ),
+            ReplicatedMapping(
+                megatron_param="decoder.layers.*.layer_scalar",
+                hf_param="model.layers.*.layer_scalar",
+            ),
+            ReplicatedMapping(
+                megatron_param="decoder.layers.*.mlp.router.per_expert_scale",
+                hf_param="model.layers.*.router.per_expert_scale",
+            ),
+            ReplicatedMapping(
+                megatron_param="decoder.layers.*.mlp.router.scale",
+                hf_param="model.layers.*.router.scale",
+            ),
+            ReplicatedMapping(
+                megatron_param="decoder.layers.*.pffl_weight",
+                hf_param="model.layers.*.pre_feedforward_layernorm.weight",
+            ),
+            ReplicatedMapping(
+                megatron_param="decoder.layers.*.mlp.post_moe_layernorm.weight",
+                hf_param="model.layers.*.post_feedforward_layernorm_2.weight",
+            ),
+            ReplicatedMapping(
+                megatron_param="decoder.layers.*.post_ffn_layernorm.weight",
+                hf_param="model.layers.*.post_feedforward_layernorm.weight",
+            ),
+        ])
+        return MegatronMappingRegistry(*mapping_list)
+
+    def _split_qkv_linear_out_weight(self, megatron_model, linear_out_weight):
+        """Detect global vs sliding layers by tensor size for LoRA export."""
+        model = megatron_model[0] if isinstance(megatron_model, list) else megatron_model
+        config = model.config
+        feature_dim = linear_out_weight.shape[-1] if linear_out_weight.ndim == 2 else None
+
+        qkv_total_sliding = config.num_attention_heads + 2 * config.num_query_groups
+        expected_numel_sliding = qkv_total_sliding * config.kv_channels * (feature_dim or 1)
+
+        if linear_out_weight.numel() != expected_numel_sliding and hasattr(config, "global_head_dim"):
+            num_kv_global = config.num_global_key_value_heads
+            head_size_global = config.global_head_dim
+
+            class _GlobalAttnCfg:
+                num_attention_heads = config.num_attention_heads
+                num_query_groups = num_kv_global
+                kv_channels = head_size_global
+                hidden_size = config.hidden_size
+                attention_output_gate = getattr(config, "attention_output_gate", False)
+
+            q_out, k_out, _ = split_qkv_weights(_GlobalAttnCfg(), linear_out_weight, feature_dim=feature_dim)
+            return {"q_proj": q_out, "k_proj": k_out, "v_proj": ABSENT_PROJECTION}
+
+        return super()._split_qkv_linear_out_weight(megatron_model, linear_out_weight)
+
+
+# ---------------------------------------------------------------------------
+# Gemma4VLBridge — VL ConditionalGeneration bridge, inherits Gemma4Bridge
+# ---------------------------------------------------------------------------
+
+
+@MegatronModelBridge.register_bridge(
+    source="Gemma4ForConditionalGeneration",
+    target=Gemma4VLModel,
+    provider=Gemma4VLModelProvider,
+    model_type="gemma4_vl",
+)
+class Gemma4VLBridge(Gemma4Bridge):
+    """Megatron Bridge for Gemma 4 Vision-Language models.
+
+    Inherits all Dense/MoE logic from Gemma4Bridge and adds VL-specific:
+    - vision_tower and embed_vision weight mappings
+    - Dense E4B VL provider construction
+    - ``GEMMA4_CONVERSION_MODE`` dispatch (text / auto / vl)
+    """
+
+    def provider_bridge(self, hf_pretrained: PreTrainedVLM) -> "Gemma4VLModelProvider | Gemma4DenseVLProvider | Gemma4DenseProvider":
+        hf_config = hf_pretrained.config
+        text_config = hf_config.text_config
+        vision_config = hf_config.vision_config
+
+        if not getattr(text_config, "enable_moe_block", False):
+            self._is_dense = True
+            if self._conversion_mode() == "text":
+                return self._build_dense_provider(text_config)
+            return self._build_dense_vl_provider(hf_config, text_config, vision_config)
+
+        self._is_dense = False
+
+        provider_kwargs = self.hf_config_to_provider_kwargs(text_config)
+        provider = Gemma4VLModelProvider(**provider_kwargs)
+
+        provider.window_size = getattr(text_config, "sliding_window", 1024)
+        provider.rotary_base = (
+            rope_local_base_freq_from_hf(text_config),
+            rope_theta_from_hf(text_config),
+        )
+
+        head_dim = getattr(text_config, "head_dim", 256)
+        provider.softmax_scale = 1.0
+        provider.kv_channels = head_dim
+        provider.qk_layernorm = True
+
+        provider.global_head_dim = getattr(text_config, "global_head_dim", 512)
+        provider.num_global_key_value_heads = getattr(text_config, "num_global_key_value_heads", 2)
+        provider.attention_k_eq_v = getattr(text_config, "attention_k_eq_v", False)
+
+        rope_params = getattr(text_config, "rope_parameters", {})
+        if isinstance(rope_params, dict):
+            full_attn_rope = rope_params.get("full_attention", {})
+            provider.global_rotary_percent = full_attn_rope.get("partial_rotary_factor", 0.25)
+
+        layer_types = getattr(text_config, "layer_types", None)
+        if layer_types:
+            provider.interleaved_attn_pattern = _infer_attn_pattern(layer_types)
+
+        if getattr(text_config, "enable_moe_block", False):
+            provider.num_moe_experts = getattr(text_config, "num_experts", None) or 128
+            provider.moe_router_topk = getattr(text_config, "top_k_experts", None) or 8
+            provider.moe_ffn_hidden_size = getattr(text_config, "moe_intermediate_size", None) or 704
+            provider.moe_shared_expert_intermediate_size = getattr(text_config, "intermediate_size", 2112)
+            provider.moe_shared_expert_overlap = False
+            provider.moe_shared_expert_gate = False
+            provider.moe_layer_freq = 1
+
+        provider.final_logit_softcapping = getattr(text_config, "final_logit_softcapping", 30.0)
+        provider.bf16 = True
+        provider.params_dtype = torch.bfloat16
+        provider.autocast_dtype = torch.bfloat16
+        provider.make_vocab_size_divisible_by = 128
+
+        provider.vision_config = vision_config
+        provider.text_config = text_config
+        provider.audio_config = getattr(hf_config, "audio_config", None)
+        provider.vision_soft_tokens_per_image = getattr(hf_config, "vision_soft_tokens_per_image", 280)
+        provider.bos_token_id = getattr(hf_config, "bos_token_id", 2)
+        provider.eos_token_id = getattr(hf_config, "eos_token_id", 1)
+        provider.image_token_id = getattr(hf_config, "image_token_id", 258_880)
+        provider.video_token_id = getattr(hf_config, "video_token_id", 258_884)
+        provider.audio_token_id = getattr(hf_config, "audio_token_id", 258_881)
+
+        return provider
+
+    def _conversion_mode(self) -> str:
+        mode = getattr(self, "gemma4_conversion_mode", None) or os.environ.get("GEMMA4_CONVERSION_MODE", "auto")
+        mode = mode.lower()
+        if mode not in {"auto", "text", "vl", "audio"}:
+            raise ValueError(f"Invalid GEMMA4_CONVERSION_MODE={mode!r}; expected auto, text, vl, or audio.")
+        # "audio" is treated as full VL+audio conversion (same as "vl"/"auto")
+        return mode
+
+    def _build_dense_vl_provider(self, hf_config, text_config, vision_config) -> Gemma4DenseVLProvider:
+        """Build a Dense VL provider by copying all Dense provider fields."""
+        from dataclasses import fields
+        text_provider = self._build_dense_provider(text_config)
+        provider = Gemma4DenseVLProvider()
+        for f in fields(Gemma4DenseProvider):
+            setattr(provider, f.name, getattr(text_provider, f.name))
+
+        provider.vision_config = vision_config
+        provider.text_config = text_config
+        provider.audio_config = getattr(hf_config, "audio_config", None)
+        provider.vision_soft_tokens_per_image = getattr(hf_config, "vision_soft_tokens_per_image", 280)
+        provider.bos_token_id = getattr(hf_config, "bos_token_id", 2)
+        provider.eos_token_id = getattr(hf_config, "eos_token_id", 1)
+        provider.image_token_id = getattr(hf_config, "image_token_id", 258_880)
+        provider.video_token_id = getattr(hf_config, "video_token_id", 258_884)
+        provider.audio_token_id = getattr(hf_config, "audio_token_id", 258_881)
+        return provider
+
+    def _text_config(self):
+        hf_config = getattr(self, "hf_config", None)
+        return getattr(hf_config, "text_config", None)
+
+    def _is_dense_e4b_config(self) -> bool:
+        if getattr(self, "_is_dense", False):
+            return True
+        text_config = self._text_config()
+        return text_config is not None and not getattr(text_config, "enable_moe_block", True)
+
+    def _hf_layer_prefix(self) -> str:
+        """VLM text weights live under ``model.language_model.*``."""
+        return "model.language_model."
+
+    def _fuse_router_weight(self, hf_param: str, hf_state_dict: Mapping[str, torch.Tensor]) -> torch.Tensor:
+        """Fuse router preprocessing — VLM prefix-aware version."""
+        proj_weight = hf_state_dict[hf_param]
+        layer_match = re.search(r"layers\.(\d+)\.", hf_param)
+        if layer_match is None:
+            return proj_weight
+        layer_idx = layer_match.group(1)
+        prefix = hf_param.rsplit("layers.", 1)[0]
+        scale_key = f"{prefix}layers.{layer_idx}.router.scale"
+        ln2_key = f"{prefix}layers.{layer_idx}.pre_feedforward_layernorm_2.weight"
+        if scale_key not in hf_state_dict or ln2_key not in hf_state_dict:
+            return proj_weight
+        router_scale = hf_state_dict[scale_key].float()
+        ln2_weight = hf_state_dict[ln2_key].float()
+        hidden_size = proj_weight.shape[-1]
+        scalar_root_size = hidden_size**-0.5
+        fusion_factor = router_scale * scalar_root_size / ln2_weight
+        fused_weight = proj_weight.float() * fusion_factor.unsqueeze(0)
+        return fused_weight.to(proj_weight.dtype)
+
+    def _fuse_shared_expert_prenorm(
+        self, hf_param: dict[str, str], hf_state_dict: Mapping[str, torch.Tensor]
+    ) -> dict[str, torch.Tensor]:
+        """Fuse pre-norm correction — VLM prefix-aware version."""
+        gate_name = hf_param["gate"]
+        layer_match = re.search(r"layers\.(\d+)\.", gate_name)
+        if layer_match is None:
+            return {role: hf_state_dict[name] for role, name in hf_param.items()}
+        layer_idx = layer_match.group(1)
+        prefix = gate_name.rsplit("layers.", 1)[0]
+        pffl_key = f"{prefix}layers.{layer_idx}.pre_feedforward_layernorm.weight"
+        pffl2_key = f"{prefix}layers.{layer_idx}.pre_feedforward_layernorm_2.weight"
+        if pffl_key not in hf_state_dict or pffl2_key not in hf_state_dict:
+            return {role: hf_state_dict[name] for role, name in hf_param.items()}
+        w_pffl = hf_state_dict[pffl_key].float()
+        w_pffl2 = hf_state_dict[pffl2_key].float()
+        correction = w_pffl / w_pffl2
+        hf_weights = {}
+        for role, name in hf_param.items():
+            weight = hf_state_dict[name]
+            fused = weight.float() * correction.unsqueeze(0)
+            hf_weights[role] = fused.to(weight.dtype)
+        return hf_weights
+
+    def maybe_modify_loaded_hf_weight(
+        self, hf_param: str | dict[str, str], hf_state_dict: Mapping[str, torch.Tensor]
+    ) -> torch.Tensor:
+        """Handle special weight loading for Gemma 4 VLM."""
         if self._is_dense_e4b_config() and isinstance(hf_param, dict) and "v" in hf_param:
             k_name = hf_param["k"]
             v_name = hf_param["v"]
@@ -284,142 +703,45 @@ class Gemma4VLBridge(MegatronModelBridge):
                 k_zero = torch.zeros(kv_shape, dtype=q_weight.dtype, device=q_weight.device)
                 return {"q": q_weight, "k": k_zero, "v": torch.zeros_like(k_zero)}
 
-        # Handle K=V on global layers
-        if isinstance(hf_param, dict) and "v" in hf_param:
-            v_name = hf_param["v"]
-            if v_name not in hf_state_dict:
-                k_name = hf_param["k"]
-                hf_weights = {}
-                for role, name in hf_param.items():
-                    if role == "v":
-                        hf_weights[role] = hf_state_dict[k_name].clone()
-                    else:
-                        hf_weights[role] = hf_state_dict[name]
-                return hf_weights
-
-        # Fuse pre-norm correction into shared expert gate/up weights
-        if isinstance(hf_param, dict) and "gate" in hf_param:
-            gate_name = hf_param["gate"]
-            if "mlp.gate_proj" in gate_name:
-                return self._fuse_shared_expert_prenorm(hf_param, hf_state_dict)
-
-        # Fuse router scaling into router.proj.weight
-        if isinstance(hf_param, str) and hf_param.endswith("router.proj.weight"):
-            return self._fuse_router_weight(hf_param, hf_state_dict)
-
         return super().maybe_modify_loaded_hf_weight(hf_param, hf_state_dict)
 
-    def _fuse_router_weight(self, hf_param: str, hf_state_dict: Mapping[str, torch.Tensor]) -> torch.Tensor:
-        """Fuse router preprocessing into projection weight (VLM version)."""
-        proj_weight = hf_state_dict[hf_param]
-
-        layer_match = re.search(r"layers\.(\d+)\.", hf_param)
-        if layer_match is None:
-            return proj_weight
-        layer_idx = layer_match.group(1)
-
-        # VLM prefix: language_model.layers.{idx}.router.*
-        prefix = hf_param.rsplit("layers.", 1)[0]
-        scale_key = f"{prefix}layers.{layer_idx}.router.scale"
-        ln2_key = f"{prefix}layers.{layer_idx}.pre_feedforward_layernorm_2.weight"
-
-        if scale_key not in hf_state_dict or ln2_key not in hf_state_dict:
-            return proj_weight
-
-        router_scale = hf_state_dict[scale_key].float()
-        ln2_weight = hf_state_dict[ln2_key].float()
-        hidden_size = proj_weight.shape[-1]
-        scalar_root_size = hidden_size**-0.5
-
-        fusion_factor = router_scale * scalar_root_size / ln2_weight
-        fused_weight = proj_weight.float() * fusion_factor.unsqueeze(0)
-        return fused_weight.to(proj_weight.dtype)
-
-    def _fuse_shared_expert_prenorm(
-        self, hf_param: dict[str, str], hf_state_dict: Mapping[str, torch.Tensor]
-    ) -> dict[str, torch.Tensor]:
-        """Fuse pre-norm correction into shared expert gate/up weights (VLM version)."""
-        gate_name = hf_param["gate"]
-        layer_match = re.search(r"layers\.(\d+)\.", gate_name)
-        if layer_match is None:
-            return {role: hf_state_dict[name] for role, name in hf_param.items()}
-
-        layer_idx = layer_match.group(1)
-        prefix = gate_name.rsplit("layers.", 1)[0]
-        pffl_key = f"{prefix}layers.{layer_idx}.pre_feedforward_layernorm.weight"
-        pffl2_key = f"{prefix}layers.{layer_idx}.pre_feedforward_layernorm_2.weight"
-
-        if pffl_key not in hf_state_dict or pffl2_key not in hf_state_dict:
-            return {role: hf_state_dict[name] for role, name in hf_param.items()}
-
-        w_pffl = hf_state_dict[pffl_key].float()
-        w_pffl2 = hf_state_dict[pffl2_key].float()
-        correction = w_pffl / w_pffl2
-
-        hf_weights = {}
-        for role, name in hf_param.items():
-            weight = hf_state_dict[name]
-            fused = weight.float() * correction.unsqueeze(0)
-            hf_weights[role] = fused.to(weight.dtype)
-        return hf_weights
-
-    def _hf_layer_prefix(self) -> str:
-        """VLM text weights live under ``model.language_model.*``."""
-        return "model.language_model."
-
-    def _text_config(self):
-        hf_config = getattr(self, "hf_config", None)
-        return getattr(hf_config, "text_config", None)
-
-    def _is_dense_e4b_config(self) -> bool:
-        if getattr(self, "_is_dense_e4b", False):
-            return True
-        text_config = self._text_config()
-        return text_config is not None and not getattr(text_config, "enable_moe_block", True)
-
-    def _is_dense_e4b_text_only(self) -> bool:
-        return getattr(self, "_is_dense_e4b_text_only", False) or self._conversion_mode() == "text"
-
     def mapping_registry(self) -> MegatronMappingRegistry:
-        """Dispatch to Dense E4B or MoE VLM mappings."""
+        """Dispatch to Dense or MoE VLM mappings."""
         if self._is_dense_e4b_config():
-            if self._is_dense_e4b_text_only():
-                return Gemma4Bridge._dense_e4b_mapping_registry(self, megatron_prefix="")
-            return self._dense_e4b_vl_mapping_registry()
+            if self._conversion_mode() == "text":
+                return self._dense_mapping_registry(megatron_prefix="")
+            return self._dense_vl_mapping_registry()
         return self._moe_vl_mapping_registry()
 
-    def _dense_e4b_vl_mapping_registry(self) -> MegatronMappingRegistry:
-        """Define parameter mappings for full Dense E4B VL checkpoints."""
-        registry = Gemma4Bridge._dense_e4b_mapping_registry(self, megatron_prefix="language_model.")
+    def _dense_vl_mapping_registry(self) -> MegatronMappingRegistry:
+        """Dense E4B VL: language mappings + vision tower + audio tower."""
+        registry = self._dense_mapping_registry(megatron_prefix="language_model.")
         mapping_list = list(registry.mappings)
-        mapping_list.extend(
-            [
-                ReplicatedMapping(
-                    megatron_param="vision_tower.**",
-                    hf_param="model.vision_tower.**",
-                ),
-                ReplicatedMapping(
-                    megatron_param="embed_vision.**",
-                    hf_param="model.embed_vision.**",
-                ),
-            ]
-        )
+        mapping_list.extend([
+            ReplicatedMapping(
+                megatron_param="vision_tower.**",
+                hf_param="model.vision_tower.**",
+            ),
+            ReplicatedMapping(
+                megatron_param="embed_vision.**",
+                hf_param="model.embed_vision.**",
+            ),
+            ReplicatedMapping(
+                megatron_param="audio_tower.**",
+                hf_param="model.audio_tower.**",
+            ),
+            ReplicatedMapping(
+                megatron_param="embed_audio.**",
+                hf_param="model.embed_audio.**",
+            ),
+        ])
         return MegatronMappingRegistry(*mapping_list)
 
     def _moe_vl_mapping_registry(self) -> MegatronMappingRegistry:
-        """Define parameter mappings for Gemma 4 MoE VLM.
-
-        HF VLM param names (raw safetensors keys include outer ``model.`` prefix):
-        - ``model.language_model.layers.*`` → language model
-        - ``model.vision_tower.*`` → vision encoder (replicated)
-        - ``model.embed_vision.*`` → multimodal projector (replicated)
-
-        """
+        """MoE VL parameter mappings."""
         param_mappings = {
-            # === Embeddings ===
             "language_model.embedding.word_embeddings.weight": "model.language_model.embed_tokens.weight",
             "language_model.decoder.final_layernorm.weight": "model.language_model.norm.weight",
-            # === Per-layer attention ===
             "language_model.decoder.layers.*.self_attention.linear_qkv.layer_norm_weight": (
                 "model.language_model.layers.*.input_layernorm.weight"
             ),
@@ -438,34 +760,28 @@ class Gemma4VLBridge(MegatronModelBridge):
             "language_model.decoder.layers.*.self_attention.linear_proj.post_layernorm.weight": (
                 "model.language_model.layers.*.post_attention_layernorm.weight"
             ),
-            # === Post-feedforward layernorm ===
             "language_model.decoder.layers.*.post_ffn_layernorm.weight": (
                 "model.language_model.layers.*.post_feedforward_layernorm.weight"
             ),
-            # === Pre-MLP layernorm (MoE pre-norm for routed experts) ===
             "language_model.decoder.layers.*.pre_mlp_layernorm.weight": (
                 "model.language_model.layers.*.pre_feedforward_layernorm_2.weight"
             ),
-            # Dense MLP → Shared Expert fc2
             "language_model.decoder.layers.*.mlp.shared_experts.linear_fc2.weight": (
                 "model.language_model.layers.*.mlp.down_proj.weight"
             ),
             "language_model.decoder.layers.*.mlp.post_shared_expert_layernorm.weight": (
                 "model.language_model.layers.*.post_feedforward_layernorm_1.weight"
             ),
-            # MoE Router
-            "language_model.decoder.layers.*.mlp.router.weight": ("model.language_model.layers.*.router.proj.weight"),
+            "language_model.decoder.layers.*.mlp.router.weight": "model.language_model.layers.*.router.proj.weight",
             "language_model.decoder.layers.*.mlp.linear_fc2.weight": (
                 "model.language_model.layers.*.mlp.down_proj.weight"
             ),
-            "language_model.decoder.layers.*.mlp.linear_fc1.layer_norm_weight": "model.language_model.layers.*.post_attention_layernorm.weight",
+            "language_model.decoder.layers.*.mlp.linear_fc1.layer_norm_weight": (
+                "model.language_model.layers.*.post_attention_layernorm.weight"
+            ),
         }
 
-        mapping_list = []
-        for megatron_param, hf_param in param_mappings.items():
-            mapping_list.append(AutoMapping(megatron_param=megatron_param, hf_param=hf_param))
-
-        # === QKV: K=V tolerant mapping ===
+        mapping_list = [AutoMapping(megatron_param=m, hf_param=h) for m, h in param_mappings.items()]
         mapping_list.append(
             _Gemma4QKVMapping(
                 megatron_param="language_model.decoder.layers.*.self_attention.linear_qkv.weight",
@@ -474,113 +790,60 @@ class Gemma4VLBridge(MegatronModelBridge):
                 v="model.language_model.layers.*.self_attn.v_proj.weight",
             )
         )
-
-        mapping_list.extend(
-            [
-                # === Dense MLP → Shared Expert gated FC1 ===
-                GatedMLPMapping(
-                    megatron_param="language_model.decoder.layers.*.mlp.shared_experts.linear_fc1.weight",
-                    gate="model.language_model.layers.*.mlp.gate_proj.weight",
-                    up="model.language_model.layers.*.mlp.up_proj.weight",
-                ),
-                # === Dense MLP ===
-                GatedMLPMapping(
-                    megatron_param="language_model.decoder.layers.*.mlp.linear_fc1.weight",
-                    gate="model.language_model.layers.*.mlp.gate_proj.weight",
-                    up="model.language_model.layers.*.mlp.up_proj.weight",
-                ),
-                # === MoE Experts (fused format) ===
-                FusedGatedExpertMapping(
-                    megatron_param="language_model.decoder.layers.*.mlp.experts.linear_fc1.weight*",
-                    hf_param="model.language_model.layers.*.experts.gate_up_proj",
-                ),
-                FusedExpertMapping(
-                    megatron_param="language_model.decoder.layers.*.mlp.experts.linear_fc2.weight*",
-                    hf_param="model.language_model.layers.*.experts.down_proj",
-                ),
-                # === Router per-expert scaling (buffer) ===
-                ReplicatedMapping(
-                    megatron_param="language_model.decoder.layers.*.mlp.router.per_expert_scale",
-                    hf_param="model.language_model.layers.*.router.per_expert_scale",
-                ),
-                # === Router input scale (fused into router weight on import; stored as buffer) ===
-                ReplicatedMapping(
-                    megatron_param="language_model.decoder.layers.*.mlp.router.scale",
-                    hf_param="model.language_model.layers.*.router.scale",
-                ),
-                # === Dense/shared-expert pre-norm (fused into gate/up on import; stored as buffer) ===
-                ReplicatedMapping(
-                    megatron_param="language_model.decoder.layers.*.pffl_weight",
-                    hf_param="model.language_model.layers.*.pre_feedforward_layernorm.weight",
-                ),
-                # === Post-MoE layernorm ===
-                ReplicatedMapping(
-                    megatron_param="language_model.decoder.layers.*.mlp.post_moe_layernorm.weight",
-                    hf_param="model.language_model.layers.*.post_feedforward_layernorm_2.weight",
-                ),
-            ]
-        )
-
-        mapping_list.extend(
-            [
-                # === Vision tower (replicated — all weights pass through) ===
-                ReplicatedMapping(
-                    megatron_param="vision_tower.**",
-                    hf_param="model.vision_tower.**",
-                ),
-                # === Multimodal embedder (replicated) ===
-                ReplicatedMapping(
-                    megatron_param="embed_vision.**",
-                    hf_param="model.embed_vision.**",
-                ),
-                # === Per-layer output scaling (buffer, common to both MoE and dense) ===
-                ReplicatedMapping(
-                    megatron_param="language_model.decoder.layers.*.layer_scalar",
-                    hf_param="model.language_model.layers.*.layer_scalar",
-                ),
-            ]
-        )
-
+        mapping_list.extend([
+            GatedMLPMapping(
+                megatron_param="language_model.decoder.layers.*.mlp.shared_experts.linear_fc1.weight",
+                gate="model.language_model.layers.*.mlp.gate_proj.weight",
+                up="model.language_model.layers.*.mlp.up_proj.weight",
+            ),
+            GatedMLPMapping(
+                megatron_param="language_model.decoder.layers.*.mlp.linear_fc1.weight",
+                gate="model.language_model.layers.*.mlp.gate_proj.weight",
+                up="model.language_model.layers.*.mlp.up_proj.weight",
+            ),
+            FusedGatedExpertMapping(
+                megatron_param="language_model.decoder.layers.*.mlp.experts.linear_fc1.weight*",
+                hf_param="model.language_model.layers.*.experts.gate_up_proj",
+            ),
+            FusedExpertMapping(
+                megatron_param="language_model.decoder.layers.*.mlp.experts.linear_fc2.weight*",
+                hf_param="model.language_model.layers.*.experts.down_proj",
+            ),
+            ReplicatedMapping(
+                megatron_param="language_model.decoder.layers.*.mlp.router.per_expert_scale",
+                hf_param="model.language_model.layers.*.router.per_expert_scale",
+            ),
+            ReplicatedMapping(
+                megatron_param="language_model.decoder.layers.*.mlp.router.scale",
+                hf_param="model.language_model.layers.*.router.scale",
+            ),
+            ReplicatedMapping(
+                megatron_param="language_model.decoder.layers.*.pffl_weight",
+                hf_param="model.language_model.layers.*.pre_feedforward_layernorm.weight",
+            ),
+            ReplicatedMapping(
+                megatron_param="language_model.decoder.layers.*.mlp.post_moe_layernorm.weight",
+                hf_param="model.language_model.layers.*.post_feedforward_layernorm_2.weight",
+            ),
+            ReplicatedMapping(
+                megatron_param="vision_tower.**",
+                hf_param="model.vision_tower.**",
+            ),
+            ReplicatedMapping(
+                megatron_param="embed_vision.**",
+                hf_param="model.embed_vision.**",
+            ),
+            ReplicatedMapping(
+                megatron_param="audio_tower.**",
+                hf_param="model.audio_tower.**",
+            ),
+            ReplicatedMapping(
+                megatron_param="embed_audio.**",
+                hf_param="model.embed_audio.**",
+            ),
+            ReplicatedMapping(
+                megatron_param="language_model.decoder.layers.*.layer_scalar",
+                hf_param="model.language_model.layers.*.layer_scalar",
+            ),
+        ])
         return MegatronMappingRegistry(*mapping_list)
-
-    def _split_qkv_linear_out_weight(self, megatron_model, linear_out_weight):
-        """Override for Gemma4 dual-attention: detect global vs sliding layers by tensor size.
-
-        Gemma4 interleaves sliding-window and full (global) attention layers with different
-        head configurations:
-          - Sliding:  kv_channels=256,         num_query_groups=num_key_value_heads
-          - Global:   global_head_dim=512,      num_global_key_value_heads=2, K=V tying
-
-        For global layers the linear_qkv LoRA output tensor is larger than the sliding
-        expectation.  We detect this and re-split using the global head dimensions.
-        For global layers ``v_proj`` is set to ``ABSENT_PROJECTION`` because HF global
-        attention has no v_proj weight (K=V tying); the export loop skips it.
-        """
-        model = megatron_model[0] if isinstance(megatron_model, list) else megatron_model
-        config = model.config
-        feature_dim = linear_out_weight.shape[-1] if linear_out_weight.ndim == 2 else None
-
-        # Expected numel for a sliding-attention layer
-        qkv_total_sliding = config.num_attention_heads + 2 * config.num_query_groups
-        expected_numel_sliding = qkv_total_sliding * config.kv_channels * (feature_dim or 1)
-
-        if linear_out_weight.numel() != expected_numel_sliding and hasattr(config, "global_head_dim"):
-            # Global attention layer — use per-layer override dimensions
-            num_kv_global = config.num_global_key_value_heads
-            head_size_global = config.global_head_dim
-
-            # Lightweight proxy: split_qkv_weights only reads these four attributes
-            class _GlobalAttnCfg:
-                num_attention_heads = config.num_attention_heads
-                num_query_groups = num_kv_global
-                kv_channels = head_size_global
-                hidden_size = config.hidden_size
-                attention_output_gate = getattr(config, "attention_output_gate", False)
-
-            q_out, k_out, _ = split_qkv_weights(_GlobalAttnCfg(), linear_out_weight, feature_dim=feature_dim)
-            # v_proj is absent in HF global attention (K=V tying).  Return ABSENT_PROJECTION
-            # so the caller knows this is intentional and not a bug (a missing key would
-            # raise KeyError; None would hit the assert).
-            return {"q_proj": q_out, "k_proj": k_out, "v_proj": ABSENT_PROJECTION}
-
-        return super()._split_qkv_linear_out_weight(megatron_model, linear_out_weight)
