+++ tests/unit_tests/models/gemma_vl/test_gemma4_vl_bridge.py
@@ -12,6 +12,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+"""Unit tests for Gemma4Bridge (CausalLM) and Gemma4VLBridge (ConditionalGeneration)."""
+
 from unittest.mock import Mock
 
 import pytest
@@ -19,25 +21,143 @@ import torch
 from transformers import GenerationConfig, SiglipVisionConfig
 
 from megatron.bridge.models.conversion.mapping_registry import MegatronMappingRegistry
-from megatron.bridge.models.gemma.gemma4_layer_specs import Gemma4E4BProvider
-from megatron.bridge.models.gemma_vl.gemma4_vl_bridge import Gemma4VLBridge
-from megatron.bridge.models.gemma_vl.gemma4_vl_provider import Gemma4E4BVLProvider, Gemma4VLModelProvider
+from megatron.bridge.models.conversion.model_bridge import MegatronModelBridge
+from megatron.bridge.models.gemma_vl.gemma4_vl_bridge import (
+    Gemma4Bridge,
+    Gemma4VLBridge,
+    _infer_attn_pattern,
+)
+from megatron.bridge.models.gemma_vl.gemma4_vl_provider import (
+    Gemma4DenseVLProvider,
+    Gemma4ModelProvider,
+    Gemma4VLModelProvider,
+)
+from megatron.bridge.models.gemma_vl.modeling_gemma4_vl import Gemma4DenseProvider
+from megatron.bridge.models.hf_pretrained.causal_lm import PreTrainedCausalLM
 from megatron.bridge.models.hf_pretrained.vlm import PreTrainedVLM
 
 
-# ---------------------------------------------------------------------------
-# Fixtures
-# ---------------------------------------------------------------------------
+# ===========================================================================
+# Shared fixtures
+# ===========================================================================
+
+
+@pytest.fixture
+def mock_vision_config():
+    config = SiglipVisionConfig()
+    config.hidden_size = 1152
+    config.intermediate_size = 4304
+    config.num_hidden_layers = 27
+    config.num_attention_heads = 16
+    config.patch_size = 14
+    config.image_size = 896
+    return config
+
+
+# ===========================================================================
+# CausalLM (Gemma4Bridge) fixtures
+# ===========================================================================
+
+
+@pytest.fixture
+def mock_hf_config_causal_moe():
+    """Flat Gemma4 CausalLM config (MoE: 26B-A4B)."""
+    cfg = Mock(spec=[])
+    cfg.num_hidden_layers = 62
+    cfg.hidden_size = 2816
+    cfg.intermediate_size = 2112
+    cfg.moe_intermediate_size = 704
+    cfg.num_attention_heads = 8
+    cfg.num_key_value_heads = 4
+    cfg.head_dim = 256
+    cfg.global_head_dim = 512
+    cfg.num_global_key_value_heads = 2
+    cfg.initializer_range = 0.02
+    cfg.rms_norm_eps = 1e-6
+    cfg.vocab_size = 262144
+    cfg.max_position_embeddings = 131072
+    cfg.sliding_window = 1024
+    cfg.rope_theta = 1000000.0
+    cfg.rope_local_base_freq = 10000.0
+    cfg.rope_parameters = {"full_attention": {"partial_rotary_factor": 0.25}}
+    cfg.query_pre_attn_scalar = 1.0
+    cfg.hidden_act = "gelu_pytorch_tanh"
+    cfg.torch_dtype = "bfloat16"
+    cfg.enable_moe_block = True
+    cfg.num_experts = 128
+    cfg.top_k_experts = 8
+    cfg.layer_types = (
+        ["sliding_attention"] * 5 + ["full_attention"] + ["sliding_attention"] * 5 + ["full_attention"]
+    )
+    cfg.final_logit_softcapping = 30.0
+    return cfg
+
+
+@pytest.fixture
+def mock_hf_config_causal_dense():
+    """Flat Gemma4 CausalLM config (Dense: enable_moe_block=False)."""
+    cfg = Mock(spec=[])
+    cfg.num_hidden_layers = 62
+    cfg.hidden_size = 2816
+    cfg.intermediate_size = 2112
+    cfg.moe_intermediate_size = 1408
+    cfg.num_attention_heads = 8
+    cfg.num_key_value_heads = 4
+    cfg.head_dim = 256
+    cfg.global_head_dim = 512
+    cfg.num_global_key_value_heads = 2
+    cfg.initializer_range = 0.02
+    cfg.rms_norm_eps = 1e-6
+    cfg.vocab_size = 262144
+    cfg.max_position_embeddings = 131072
+    cfg.sliding_window = 1024
+    cfg.rope_theta = 1000000.0
+    cfg.rope_local_base_freq = 10000.0
+    cfg.rope_parameters = {"full_attention": {"partial_rotary_factor": 0.25}}
+    cfg.query_pre_attn_scalar = 1.0
+    cfg.hidden_act = "gelu_pytorch_tanh"
+    cfg.torch_dtype = "bfloat16"
+    cfg.enable_moe_block = False
+    cfg.num_experts = 256
+    cfg.top_k_experts = 16
+    cfg.layer_types = (
+        ["sliding_attention"] * 5 + ["full_attention"] + ["sliding_attention"] * 5 + ["full_attention"]
+    )
+    cfg.final_logit_softcapping = 30.0
+    return cfg
+
+
+@pytest.fixture
+def mock_causal_pretrained(mock_hf_config_causal_moe):
+    p = Mock(spec=PreTrainedCausalLM)
+    p.config = mock_hf_config_causal_moe
+    return p
+
+
+@pytest.fixture
+def mock_causal_dense_pretrained(mock_hf_config_causal_dense):
+    p = Mock(spec=PreTrainedCausalLM)
+    p.config = mock_hf_config_causal_dense
+    return p
+
+
+@pytest.fixture
+def causal_bridge():
+    return Gemma4Bridge()
+
+
+# ===========================================================================
+# VL (Gemma4VLBridge) fixtures
+# ===========================================================================
 
 
 @pytest.fixture
 def mock_text_config_moe():
-    """Mock text config for Gemma 4 26B-A4B (MoE model)."""
     config = Mock(spec=[])
     config.num_hidden_layers = 62
     config.hidden_size = 2816
-    config.intermediate_size = 2112  # shared expert FFN size
-    config.moe_intermediate_size = 704  # routed expert FFN size
+    config.intermediate_size = 2112
+    config.moe_intermediate_size = 704
     config.num_attention_heads = 8
     config.num_key_value_heads = 4
     config.head_dim = 256
@@ -49,17 +169,15 @@ def mock_text_config_moe():
     config.max_position_embeddings = 131072
     config.sliding_window = 1024
     config.rope_theta = 1000000.0
-    config.query_pre_attn_scalar = 1.0  # not used for scale (softmax_scale=1.0)
+    config.query_pre_attn_scalar = 1.0
     config.rope_scaling = None
     config.rope_local_base_freq = 10000.0
     config.rope_parameters = {"rope_local_base_freq": 10000.0}
     config.hidden_act = "gelu_pytorch_tanh"
     config.torch_dtype = "bfloat16"
-    # MoE fields
     config.enable_moe_block = True
     config.num_experts = 128
     config.top_k_experts = 8
-    # Attention pattern
     config.layer_types = (
         ["sliding_attention"] * 5 + ["full_attention"] + ["sliding_attention"] * 5 + ["full_attention"]
     )
@@ -68,15 +186,35 @@ def mock_text_config_moe():
 
 
 @pytest.fixture
-def mock_vision_config():
-    """Mock vision config for Gemma 4 VL."""
-    config = SiglipVisionConfig()
-    config.hidden_size = 1152
-    config.intermediate_size = 4304
-    config.num_hidden_layers = 27
-    config.num_attention_heads = 16
-    config.patch_size = 14
-    config.image_size = 896
+def mock_text_config_dense():
+    config = Mock(spec=[])
+    config.num_hidden_layers = 62
+    config.hidden_size = 2816
+    config.intermediate_size = 2112
+    config.moe_intermediate_size = 704
+    config.num_attention_heads = 8
+    config.num_key_value_heads = 4
+    config.head_dim = 256
+    config.global_head_dim = 512
+    config.num_global_key_value_heads = 2
+    config.initializer_range = 0.02
+    config.rms_norm_eps = 1e-6
+    config.vocab_size = 262144
+    config.max_position_embeddings = 131072
+    config.sliding_window = 1024
+    config.rope_theta = 1000000.0
+    config.query_pre_attn_scalar = 1.0
+    config.rope_scaling = None
+    config.rope_local_base_freq = 10000.0
+    config.rope_parameters = {"rope_local_base_freq": 10000.0}
+    config.hidden_act = "gelu_pytorch_tanh"
+    config.torch_dtype = "bfloat16"
+    config.hidden_size_per_layer_input = 0
+    config.enable_moe_block = False
+    config.layer_types = (
+        ["sliding_attention"] * 5 + ["full_attention"] + ["sliding_attention"] * 5 + ["full_attention"]
+    )
+    config.final_logit_softcapping = 30.0
     return config
 
 
@@ -94,51 +232,6 @@ def mock_hf_config_moe(mock_text_config_moe, mock_vision_config):
 
 
 @pytest.fixture
-def mock_hf_pretrained_moe(mock_hf_config_moe):
-    pretrained = Mock(spec=PreTrainedVLM)
-    pretrained.config = mock_hf_config_moe
-    pretrained.generation_config = GenerationConfig()
-    return pretrained
-
-
-@pytest.fixture
-def mock_text_config_dense():
-    """Mock text config for Gemma 4 26B-A4B (MoE model)."""
-    config = Mock(spec=[])
-    config.num_hidden_layers = 62
-    config.hidden_size = 2816
-    config.intermediate_size = 2112  # shared expert FFN size
-    config.moe_intermediate_size = 704  # routed expert FFN size
-    config.num_attention_heads = 8
-    config.num_key_value_heads = 4
-    config.head_dim = 256
-    config.global_head_dim = 512
-    config.num_global_key_value_heads = 2
-    config.initializer_range = 0.02
-    config.rms_norm_eps = 1e-6
-    config.vocab_size = 262144
-    config.max_position_embeddings = 131072
-    config.sliding_window = 1024
-    config.rope_theta = 1000000.0
-    config.query_pre_attn_scalar = 1.0  # not used for scale (softmax_scale=1.0)
-    config.rope_scaling = None
-    config.rope_local_base_freq = 10000.0
-    config.rope_parameters = {"rope_local_base_freq": 10000.0}
-    config.hidden_act = "gelu_pytorch_tanh"
-    config.torch_dtype = "bfloat16"
-    config.hidden_size_per_layer_input = 0
-    # MoE fields
-    config.enable_moe_block = False
-
-    # Attention pattern
-    config.layer_types = (
-        ["sliding_attention"] * 5 + ["full_attention"] + ["sliding_attention"] * 5 + ["full_attention"]
-    )
-    config.final_logit_softcapping = 30.0
-    return config
-
-
-@pytest.fixture
 def mock_hf_config_dense(mock_text_config_dense, mock_vision_config):
     config = Mock()
     config.text_config = mock_text_config_dense
@@ -152,137 +245,314 @@ def mock_hf_config_dense(mock_text_config_dense, mock_vision_config):
 
 
 @pytest.fixture
-def mock_hf_pretrained_dense(mock_hf_config_dense):
-    pretrained = Mock(spec=PreTrainedVLM)
-    pretrained.config = mock_hf_config_dense
-    pretrained.generation_config = GenerationConfig()
-    return pretrained
+def mock_hf_pretrained_moe(mock_hf_config_moe):
+    p = Mock(spec=PreTrainedVLM)
+    p.config = mock_hf_config_moe
+    p.generation_config = GenerationConfig()
+    return p
 
 
 @pytest.fixture
-def bridge():
+def mock_hf_pretrained_dense(mock_hf_config_dense):
+    p = Mock(spec=PreTrainedVLM)
+    p.config = mock_hf_config_dense
+    p.generation_config = GenerationConfig()
+    return p
+
+
+@pytest.fixture
+def vl_bridge():
     return Gemma4VLBridge()
 
 
-# ---------------------------------------------------------------------------
-# Initialization
-# ---------------------------------------------------------------------------
+# ===========================================================================
+# Gemma4Bridge (CausalLM) tests
+# ===========================================================================
 
 
-class TestGemma4VLBridgeInitialization:
-    def test_bridge_initialization(self, bridge):
-        assert isinstance(bridge, Gemma4VLBridge)
+class TestGemma4BridgeRegistration:
+    def test_is_subclass_of_model_bridge(self):
+        assert issubclass(Gemma4Bridge, MegatronModelBridge)
 
-    def test_bridge_has_required_methods(self, bridge):
-        assert callable(getattr(bridge, "provider_bridge", None))
-        assert callable(getattr(bridge, "mapping_registry", None))
+    def test_vl_bridge_inherits_causal_bridge(self):
+        assert issubclass(Gemma4VLBridge, Gemma4Bridge)
 
+    def test_initialization(self, causal_bridge):
+        assert isinstance(causal_bridge, Gemma4Bridge)
 
-# ---------------------------------------------------------------------------
-# provider_bridge — MoE model
-# ---------------------------------------------------------------------------
-
-
-class TestGemma4VLBridgeProviderBridgeMoE:
-    def test_returns_provider(self, bridge, mock_hf_pretrained_moe):
-        provider = bridge.provider_bridge(mock_hf_pretrained_moe)
-        assert isinstance(provider, Gemma4VLModelProvider)
-
-    def test_basic_transformer_config(self, bridge, mock_hf_pretrained_moe):
-        provider = bridge.provider_bridge(mock_hf_pretrained_moe)
-        assert provider.num_layers == 62
-        assert provider.hidden_size == 2816
-        assert provider.num_attention_heads == 8
-        assert provider.num_query_groups == 4
-        assert provider.kv_channels == 256
-        assert provider.init_method_std == 0.02
-        assert provider.layernorm_epsilon == 1e-6
-        assert provider.vocab_size == 262144
-        assert provider.seq_length == 131072
-        assert provider.window_size == 1024
-
-    def test_moe_config(self, bridge, mock_hf_pretrained_moe):
-        provider = bridge.provider_bridge(mock_hf_pretrained_moe)
-        assert provider.num_moe_experts == 128
-        assert provider.moe_router_topk == 8
-        assert provider.moe_ffn_hidden_size == 704
-        assert provider.moe_shared_expert_intermediate_size == 2112
-        assert provider.moe_layer_freq == 1
-
-    def test_softmax_scale_is_one(self, bridge, mock_hf_pretrained_moe):
-        provider = bridge.provider_bridge(mock_hf_pretrained_moe)
-        assert provider.softmax_scale == 1.0
-
-    def test_vl_specific_config(self, bridge, mock_hf_pretrained_moe):
-        provider = bridge.provider_bridge(mock_hf_pretrained_moe)
-        assert provider.image_token_id == 258_880
-        assert provider.video_token_id == 258_884
-        assert provider.bos_token_id == 2
-        assert provider.eos_token_id == 1
-        assert provider.vision_soft_tokens_per_image == 280
-
-    def test_dtype_is_bf16(self, bridge, mock_hf_pretrained_moe):
-        provider = bridge.provider_bridge(mock_hf_pretrained_moe)
-        assert provider.bf16 is True
-        assert provider.params_dtype == torch.bfloat16
-
-    def test_global_head_config(self, bridge, mock_hf_pretrained_moe):
-        provider = bridge.provider_bridge(mock_hf_pretrained_moe)
-        assert provider.global_head_dim == 512
-        assert provider.num_global_key_value_heads == 2
-
-    def test_qk_layernorm_enabled(self, bridge, mock_hf_pretrained_moe):
-        provider = bridge.provider_bridge(mock_hf_pretrained_moe)
-        assert provider.qk_layernorm is True
-
-    def test_logit_softcapping(self, bridge, mock_hf_pretrained_moe):
-        provider = bridge.provider_bridge(mock_hf_pretrained_moe)
-        assert provider.final_logit_softcapping == 30.0
-
-    def test_vision_config_set(self, bridge, mock_hf_pretrained_moe):
-        provider = bridge.provider_bridge(mock_hf_pretrained_moe)
-        assert provider.vision_config is mock_hf_pretrained_moe.config.vision_config
-        assert provider.text_config is mock_hf_pretrained_moe.config.text_config
+    def test_has_required_methods(self, causal_bridge):
+        assert callable(getattr(causal_bridge, "provider_bridge", None))
+        assert callable(getattr(causal_bridge, "mapping_registry", None))
+        assert callable(getattr(causal_bridge, "maybe_modify_loaded_hf_weight", None))
+        assert callable(getattr(causal_bridge, "maybe_modify_converted_hf_weight", None))
 
 
-# ---------------------------------------------------------------------------
-# provider_bridge — dense model
-# ---------------------------------------------------------------------------
+class TestGemma4BridgeProviderBridgeMoE:
+    """Gemma4Bridge.provider_bridge for MoE CausalLM."""
+
+    def test_returns_provider_instance(self, causal_bridge, mock_causal_pretrained):
+        provider = causal_bridge.provider_bridge(mock_causal_pretrained)
+        assert isinstance(provider, Gemma4ModelProvider)
+
+    def test_basic_transformer_config(self, causal_bridge, mock_causal_pretrained):
+        p = causal_bridge.provider_bridge(mock_causal_pretrained)
+        assert p.num_layers == 62
+        assert p.hidden_size == 2816
+        assert p.num_attention_heads == 8
+        assert p.num_query_groups == 4
+        assert p.kv_channels == 256
+        assert p.vocab_size == 262144
+        assert p.seq_length == 131072
+        assert p.init_method_std == 0.02
+        assert p.layernorm_epsilon == 1e-6
+
+    def test_moe_config(self, causal_bridge, mock_causal_pretrained):
+        p = causal_bridge.provider_bridge(mock_causal_pretrained)
+        assert p.num_moe_experts == 128
+        assert p.moe_router_topk == 8
+        assert p.moe_ffn_hidden_size == 704
+        assert p.moe_shared_expert_intermediate_size == 2112
+        assert p.moe_layer_freq == 1
+        assert p.moe_shared_expert_overlap is False
+        assert p.moe_shared_expert_gate is False
+
+    def test_window_size(self, causal_bridge, mock_causal_pretrained):
+        assert causal_bridge.provider_bridge(mock_causal_pretrained).window_size == 1024
+
+    def test_rotary_base_tuple(self, causal_bridge, mock_causal_pretrained):
+        rb = causal_bridge.provider_bridge(mock_causal_pretrained).rotary_base
+        assert isinstance(rb, tuple) and len(rb) == 2
+        assert rb[0] == 10000.0
+        assert rb[1] == 1000000.0
+
+    def test_softmax_scale_is_one(self, causal_bridge, mock_causal_pretrained):
+        assert causal_bridge.provider_bridge(mock_causal_pretrained).softmax_scale == 1.0
+
+    def test_qk_layernorm_enabled(self, causal_bridge, mock_causal_pretrained):
+        assert causal_bridge.provider_bridge(mock_causal_pretrained).qk_layernorm is True
+
+    def test_global_attention_config(self, causal_bridge, mock_causal_pretrained):
+        p = causal_bridge.provider_bridge(mock_causal_pretrained)
+        assert p.global_head_dim == 512
+        assert p.num_global_key_value_heads == 2
+        assert p.global_rotary_percent == 0.25
+
+    def test_interleaved_attn_pattern(self, causal_bridge, mock_causal_pretrained):
+        assert causal_bridge.provider_bridge(mock_causal_pretrained).interleaved_attn_pattern == (5, 1)
+
+    def test_logit_softcapping(self, causal_bridge, mock_causal_pretrained):
+        assert causal_bridge.provider_bridge(mock_causal_pretrained).final_logit_softcapping == 30.0
+
+    def test_dtype_is_bf16(self, causal_bridge, mock_causal_pretrained):
+        p = causal_bridge.provider_bridge(mock_causal_pretrained)
+        assert p.bf16 is True
+        assert p.params_dtype == torch.bfloat16
+
+    def test_different_hidden_sizes(self, causal_bridge, mock_causal_pretrained):
+        for hs in [2048, 2816, 4096]:
+            mock_causal_pretrained.config.hidden_size = hs
+            assert causal_bridge.provider_bridge(mock_causal_pretrained).hidden_size == hs
+
+    def test_different_layer_counts(self, causal_bridge, mock_causal_pretrained):
+        for nl in [32, 46, 62]:
+            mock_causal_pretrained.config.num_hidden_layers = nl
+            assert causal_bridge.provider_bridge(mock_causal_pretrained).num_layers == nl
+
+    def test_vocab_size_variants(self, causal_bridge, mock_causal_pretrained):
+        for vs in [256000, 262144, 300000]:
+            mock_causal_pretrained.config.vocab_size = vs
+            assert causal_bridge.provider_bridge(mock_causal_pretrained).vocab_size == vs
 
 
-class TestGemma4VLBridgeProviderBridgeDense:
-    def test_accepts_dense_with_hidden_size_per_layer_model(self, bridge, mock_hf_pretrained_dense):
-        """Dense E4B with per-layer inputs is supported by Gemma4E4BVLProvider."""
-        mock_hf_pretrained_dense.config.text_config.hidden_size_per_layer_input = 256
-        provider = bridge.provider_bridge(mock_hf_pretrained_dense)
-        assert isinstance(provider, Gemma4E4BVLProvider)
-        assert provider.per_layer_embed_dim == 256
+class TestGemma4BridgeProviderBridgeDense:
+    """Gemma4Bridge.provider_bridge for Dense CausalLM (enable_moe_block=False)."""
 
-    def test_returns_provider(self, bridge, mock_hf_pretrained_dense):
-        provider = bridge.provider_bridge(mock_hf_pretrained_dense)
-        assert isinstance(provider, Gemma4E4BVLProvider)
+    def test_returns_dense_provider(self, causal_bridge, mock_causal_dense_pretrained):
+        p = causal_bridge.provider_bridge(mock_causal_dense_pretrained)
+        assert isinstance(p, Gemma4DenseProvider)
 
-    def test_text_conversion_mode_returns_text_provider(self, bridge, mock_hf_pretrained_dense, monkeypatch):
-        monkeypatch.setenv("GEMMA4_CONVERSION_MODE", "text")
-        provider = bridge.provider_bridge(mock_hf_pretrained_dense)
-        assert isinstance(provider, Gemma4E4BProvider)
-        assert not isinstance(provider, Gemma4E4BVLProvider)
+    def test_basic_config_preserved(self, causal_bridge, mock_causal_dense_pretrained):
+        p = causal_bridge.provider_bridge(mock_causal_dense_pretrained)
+        assert p.num_layers == 62
+        assert p.hidden_size == 2816
+        assert p.num_attention_heads == 8
+        assert p.num_query_groups == 4
+        assert p.vocab_size == 262144
+
+    def test_does_not_copy_moe_intermediate_size(self, causal_bridge, mock_causal_dense_pretrained):
+        """Dense provider should NOT use moe_intermediate_size from HF config."""
+        p = causal_bridge.provider_bridge(mock_causal_dense_pretrained)
+        # Dense provider has its own moe_ffn_hidden_size default (704), not 1408 from HF config
+        assert p.moe_ffn_hidden_size == mock_causal_dense_pretrained.config.moe_ffn_hidden_size if hasattr(
+            mock_causal_dense_pretrained.config, "moe_ffn_hidden_size"
+        ) else True  # default kept
 
 
-# ---------------------------------------------------------------------------
-# mapping_registry
-# ---------------------------------------------------------------------------
+class TestInferAttnPattern:
+    def test_5_sliding_1_global(self):
+        lt = ["sliding_attention"] * 5 + ["full_attention"] + ["sliding_attention"] * 5 + ["full_attention"]
+        assert _infer_attn_pattern(lt) == (5, 1)
+
+    def test_all_sliding(self):
+        assert _infer_attn_pattern(["sliding_attention"] * 8) == (8, 0)
+
+    def test_single_sliding_then_global(self):
+        assert _infer_attn_pattern(["sliding_attention", "full_attention", "sliding_attention"]) == (1, 1)
+
+    def test_consecutive_global_layers(self):
+        lt = ["sliding_attention"] * 3 + ["full_attention", "full_attention"]
+        assert _infer_attn_pattern(lt) == (3, 2)
+
+    def test_global_at_start(self):
+        assert _infer_attn_pattern(["full_attention"] + ["sliding_attention"] * 5) == (0, 1)
 
 
-class TestGemma4VLBridgeMappingRegistry:
-    def test_returns_registry(self, bridge):
-        registry = bridge.mapping_registry()
-        assert isinstance(registry, MegatronMappingRegistry)
+class TestMaybeModifyLoadedHFWeightCausal:
+    """Weight modification during HF → Megatron loading (CausalLM path)."""
 
-    def test_has_mappings(self, bridge):
-        registry = bridge.mapping_registry()
-        assert len(registry.mappings) > 0
+    def _make_sd(self, layer_idx=0, hidden=8, num_experts=4):
+        p = f"model.layers.{layer_idx}"
+        sd = {
+            f"{p}.self_attn.q_proj.weight": torch.randn(hidden, hidden),
+            f"{p}.self_attn.k_proj.weight": torch.randn(hidden // 2, hidden),
+            f"{p}.router.proj.weight": torch.randn(num_experts, hidden),
+            f"{p}.router.scale": torch.ones(hidden),
+            f"{p}.pre_feedforward_layernorm_2.weight": torch.ones(hidden) * 2.0,
+            f"{p}.mlp.gate_proj.weight": torch.randn(16, hidden),
+            f"{p}.mlp.up_proj.weight": torch.randn(16, hidden),
+            f"{p}.pre_feedforward_layernorm.weight": torch.ones(hidden) * 3.0,
+        }
+        return sd
 
+    def test_kv_synthesis_when_v_proj_absent(self, causal_bridge):
+        sd = self._make_sd()
+        hf_param = {
+            "q": "model.layers.0.self_attn.q_proj.weight",
+            "k": "model.layers.0.self_attn.k_proj.weight",
+            "v": "model.layers.0.self_attn.v_proj.weight",
+        }
+        result = causal_bridge.maybe_modify_loaded_hf_weight(hf_param, sd)
+        assert isinstance(result, dict)
+        torch.testing.assert_close(result["v"], result["k"])
+
+    def test_kv_no_synthesis_when_v_present(self, causal_bridge):
+        sd = self._make_sd()
+        sd["model.layers.0.self_attn.v_proj.weight"] = torch.randn(4, 8)
+        hf_param = {
+            "q": "model.layers.0.self_attn.q_proj.weight",
+            "k": "model.layers.0.self_attn.k_proj.weight",
+            "v": "model.layers.0.self_attn.v_proj.weight",
+        }
+        result = causal_bridge.maybe_modify_loaded_hf_weight(hf_param, sd)
+        assert result is not None
+
+    def test_router_weight_fusion(self, causal_bridge):
+        hidden = 8
+        sd = self._make_sd(hidden=hidden)
+        hf_param = "model.layers.0.router.proj.weight"
+        result = causal_bridge.maybe_modify_loaded_hf_weight(hf_param, sd)
+        assert isinstance(result, torch.Tensor)
+        assert result.shape == sd[hf_param].shape
+        expected_factor = 1.0 * (hidden**-0.5) / 2.0
+        expected = (sd[hf_param].float() * expected_factor).to(sd[hf_param].dtype)
+        torch.testing.assert_close(result, expected)
+
+    def test_router_fusion_missing_keys_passthrough(self, causal_bridge):
+        sd = {"model.layers.0.router.proj.weight": torch.randn(4, 8)}
+        result = causal_bridge.maybe_modify_loaded_hf_weight("model.layers.0.router.proj.weight", sd)
+        torch.testing.assert_close(result, sd["model.layers.0.router.proj.weight"])
+
+    def test_shared_expert_prenorm_fusion(self, causal_bridge):
+        hidden = 8
+        sd = self._make_sd(hidden=hidden)
+        hf_param = {
+            "gate": "model.layers.0.mlp.gate_proj.weight",
+            "up": "model.layers.0.mlp.up_proj.weight",
+        }
+        result = causal_bridge.maybe_modify_loaded_hf_weight(hf_param, sd)
+        assert isinstance(result, dict)
+        correction = 3.0 / 2.0
+        expected = (sd["model.layers.0.mlp.gate_proj.weight"].float() * correction).to(
+            sd["model.layers.0.mlp.gate_proj.weight"].dtype
+        )
+        torch.testing.assert_close(result["gate"], expected)
+
+    def test_shared_expert_fusion_missing_keys_passthrough(self, causal_bridge):
+        sd = {
+            "model.layers.0.mlp.gate_proj.weight": torch.randn(4, 8),
+            "model.layers.0.mlp.up_proj.weight": torch.randn(4, 8),
+        }
+        hf_param = {"gate": "model.layers.0.mlp.gate_proj.weight", "up": "model.layers.0.mlp.up_proj.weight"}
+        result = causal_bridge.maybe_modify_loaded_hf_weight(hf_param, sd)
+        torch.testing.assert_close(result["gate"], sd["model.layers.0.mlp.gate_proj.weight"])
+
+
+class TestMaybeModifyConvertedHFWeightCausal:
+    """Weight un-fusion during Megatron → HF export (CausalLM path)."""
+
+    def _make_ref_sd(self, layer_idx=0, hidden=8, num_experts=4):
+        p = f"model.layers.{layer_idx}"
+        return {
+            f"{p}.router.proj.weight": torch.randn(num_experts, hidden),
+            f"{p}.router.scale": torch.ones(hidden),
+            f"{p}.pre_feedforward_layernorm_2.weight": torch.ones(hidden) * 2.0,
+            f"{p}.mlp.gate_proj.weight": torch.randn(16, hidden),
+            f"{p}.mlp.up_proj.weight": torch.randn(16, hidden),
+            f"{p}.pre_feedforward_layernorm.weight": torch.ones(hidden) * 3.0,
+        }
+
+    def test_drops_synthesized_v_proj(self, causal_bridge):
+        hf_sd = {"model.layers.0.self_attn.q_proj.weight": torch.randn(8, 8)}
+        converted = {
+            "model.layers.0.self_attn.q_proj.weight": torch.randn(8, 8),
+            "model.layers.0.self_attn.v_proj.weight": torch.randn(4, 8),
+        }
+        result = causal_bridge.maybe_modify_converted_hf_weight(None, converted, hf_sd)
+        assert "model.layers.0.self_attn.v_proj.weight" not in result
+        assert "model.layers.0.self_attn.q_proj.weight" in result
+
+    def test_router_weight_unfusion(self, causal_bridge):
+        hidden = 8
+        ref_sd = self._make_ref_sd(hidden=hidden)
+        factor = 1.0 * (hidden**-0.5) / 2.0
+        fused = (ref_sd["model.layers.0.router.proj.weight"].float() * factor).to(
+            ref_sd["model.layers.0.router.proj.weight"].dtype
+        )
+        result = causal_bridge.maybe_modify_converted_hf_weight(None, {"model.layers.0.router.proj.weight": fused}, ref_sd)
+        torch.testing.assert_close(
+            result["model.layers.0.router.proj.weight"],
+            ref_sd["model.layers.0.router.proj.weight"],
+            atol=1e-5, rtol=1e-5,
+        )
+
+    def test_shared_expert_gate_unfusion(self, causal_bridge):
+        hidden = 8
+        ref_sd = self._make_ref_sd(hidden=hidden)
+        correction = 3.0 / 2.0
+        fused = (ref_sd["model.layers.0.mlp.gate_proj.weight"].float() * correction).to(
+            ref_sd["model.layers.0.mlp.gate_proj.weight"].dtype
+        )
+        result = causal_bridge.maybe_modify_converted_hf_weight(
+            None, {"model.layers.0.mlp.gate_proj.weight": fused}, ref_sd
+        )
+        torch.testing.assert_close(
+            result["model.layers.0.mlp.gate_proj.weight"],
+            ref_sd["model.layers.0.mlp.gate_proj.weight"],
+            atol=1e-5, rtol=1e-5,
+        )
+
+    def test_empty_hf_state_dict_passthrough(self, causal_bridge):
+        converted = {"some.weight": torch.randn(4, 4)}
+        result = causal_bridge.maybe_modify_converted_hf_weight(None, converted, {})
+        assert result is converted
+
+    def test_none_hf_state_dict_passthrough(self, causal_bridge):
+        converted = {"some.weight": torch.randn(4, 4)}
+        result = causal_bridge.maybe_modify_converted_hf_weight(None, converted, None)
+        assert result is converted
+
+
+class TestGemma4BridgeMappingRegistryCausal:
     def _collect_names(self, registry):
         names = []
         for m in registry.mappings:
@@ -294,6 +564,163 @@ class TestGemma4VLBridgeMappingRegistry:
             elif isinstance(hf, str):
                 names.append(hf)
         return names
+
+    def test_returns_registry(self, causal_bridge):
+        assert isinstance(causal_bridge.mapping_registry(), MegatronMappingRegistry)
+
+    def test_has_mappings(self, causal_bridge):
+        assert len(causal_bridge.mapping_registry().mappings) > 0
+
+    def test_has_embeddings_mapping(self, causal_bridge):
+        names = self._collect_names(causal_bridge.mapping_registry())
+        assert any("embed_tokens" in n or "word_embeddings" in n for n in names)
+
+    def test_has_final_norm_mapping(self, causal_bridge):
+        names = self._collect_names(causal_bridge.mapping_registry())
+        assert any("norm" in n for n in names)
+
+    def test_has_qkv_mapping(self, causal_bridge):
+        names = self._collect_names(causal_bridge.mapping_registry())
+        assert any("linear_qkv" in n for n in names)
+
+    def test_has_router_mapping(self, causal_bridge):
+        names = self._collect_names(causal_bridge.mapping_registry())
+        assert any("router" in n for n in names)
+
+    def test_has_shared_expert_mapping(self, causal_bridge):
+        names = self._collect_names(causal_bridge.mapping_registry())
+        assert any("shared_experts" in n for n in names)
+
+    def test_has_post_moe_layernorm(self, causal_bridge):
+        names = self._collect_names(causal_bridge.mapping_registry())
+        assert any("post_moe_layernorm" in n for n in names)
+
+    def test_has_layer_scalar_mapping(self, causal_bridge):
+        names = self._collect_names(causal_bridge.mapping_registry())
+        assert any("layer_scalar" in n for n in names)
+
+    def test_uses_causal_lm_prefix(self, causal_bridge):
+        """CausalLM bridge uses model.layers.* (not model.language_model.layers.*)."""
+        names = self._collect_names(causal_bridge.mapping_registry())
+        hf_names = [n for n in names if "layers" in n]
+        assert all("language_model" not in n for n in hf_names)
+
+
+# ===========================================================================
+# Gemma4VLBridge (ConditionalGeneration) tests
+# ===========================================================================
+
+
+@pytest.fixture
+def bridge():
+    return Gemma4VLBridge()
+
+
+class TestGemma4VLBridgeInitialization:
+    def test_bridge_initialization(self, bridge):
+        assert isinstance(bridge, Gemma4VLBridge)
+
+    def test_inherits_causal_bridge(self):
+        assert issubclass(Gemma4VLBridge, Gemma4Bridge)
+
+    def test_bridge_has_required_methods(self, bridge):
+        assert callable(getattr(bridge, "provider_bridge", None))
+        assert callable(getattr(bridge, "mapping_registry", None))
+
+
+class TestGemma4VLBridgeProviderBridgeMoE:
+    def test_returns_provider(self, bridge, mock_hf_pretrained_moe):
+        assert isinstance(bridge.provider_bridge(mock_hf_pretrained_moe), Gemma4VLModelProvider)
+
+    def test_basic_transformer_config(self, bridge, mock_hf_pretrained_moe):
+        p = bridge.provider_bridge(mock_hf_pretrained_moe)
+        assert p.num_layers == 62
+        assert p.hidden_size == 2816
+        assert p.num_attention_heads == 8
+        assert p.num_query_groups == 4
+        assert p.kv_channels == 256
+        assert p.init_method_std == 0.02
+        assert p.layernorm_epsilon == 1e-6
+        assert p.vocab_size == 262144
+        assert p.seq_length == 131072
+        assert p.window_size == 1024
+
+    def test_moe_config(self, bridge, mock_hf_pretrained_moe):
+        p = bridge.provider_bridge(mock_hf_pretrained_moe)
+        assert p.num_moe_experts == 128
+        assert p.moe_router_topk == 8
+        assert p.moe_ffn_hidden_size == 704
+        assert p.moe_shared_expert_intermediate_size == 2112
+        assert p.moe_layer_freq == 1
+
+    def test_softmax_scale_is_one(self, bridge, mock_hf_pretrained_moe):
+        assert bridge.provider_bridge(mock_hf_pretrained_moe).softmax_scale == 1.0
+
+    def test_vl_specific_config(self, bridge, mock_hf_pretrained_moe):
+        p = bridge.provider_bridge(mock_hf_pretrained_moe)
+        assert p.image_token_id == 258_880
+        assert p.video_token_id == 258_884
+        assert p.bos_token_id == 2
+        assert p.eos_token_id == 1
+        assert p.vision_soft_tokens_per_image == 280
+
+    def test_dtype_is_bf16(self, bridge, mock_hf_pretrained_moe):
+        p = bridge.provider_bridge(mock_hf_pretrained_moe)
+        assert p.bf16 is True
+        assert p.params_dtype == torch.bfloat16
+
+    def test_global_head_config(self, bridge, mock_hf_pretrained_moe):
+        p = bridge.provider_bridge(mock_hf_pretrained_moe)
+        assert p.global_head_dim == 512
+        assert p.num_global_key_value_heads == 2
+
+    def test_qk_layernorm_enabled(self, bridge, mock_hf_pretrained_moe):
+        assert bridge.provider_bridge(mock_hf_pretrained_moe).qk_layernorm is True
+
+    def test_logit_softcapping(self, bridge, mock_hf_pretrained_moe):
+        assert bridge.provider_bridge(mock_hf_pretrained_moe).final_logit_softcapping == 30.0
+
+    def test_vision_config_set(self, bridge, mock_hf_pretrained_moe):
+        p = bridge.provider_bridge(mock_hf_pretrained_moe)
+        assert p.vision_config is mock_hf_pretrained_moe.config.vision_config
+        assert p.text_config is mock_hf_pretrained_moe.config.text_config
+
+
+class TestGemma4VLBridgeProviderBridgeDense:
+    def test_accepts_dense_with_per_layer_inputs(self, bridge, mock_hf_pretrained_dense):
+        mock_hf_pretrained_dense.config.text_config.hidden_size_per_layer_input = 256
+        p = bridge.provider_bridge(mock_hf_pretrained_dense)
+        assert isinstance(p, Gemma4DenseVLProvider)
+        assert p.per_layer_embed_dim == 256
+
+    def test_returns_dense_vl_provider(self, bridge, mock_hf_pretrained_dense):
+        assert isinstance(bridge.provider_bridge(mock_hf_pretrained_dense), Gemma4DenseVLProvider)
+
+    def test_text_mode_returns_text_provider(self, bridge, mock_hf_pretrained_dense, monkeypatch):
+        monkeypatch.setenv("GEMMA4_CONVERSION_MODE", "text")
+        p = bridge.provider_bridge(mock_hf_pretrained_dense)
+        assert isinstance(p, Gemma4DenseProvider)
+        assert not isinstance(p, Gemma4DenseVLProvider)
+
+
+class TestGemma4VLBridgeMappingRegistry:
+    def _collect_names(self, registry):
+        names = []
+        for m in registry.mappings:
+            if hasattr(m, "megatron_param"):
+                names.append(str(m.megatron_param))
+            hf = getattr(m, "hf_param", None)
+            if isinstance(hf, dict):
+                names.extend(str(v) for v in hf.values())
+            elif isinstance(hf, str):
+                names.append(hf)
+        return names
+
+    def test_returns_registry(self, bridge):
+        assert isinstance(bridge.mapping_registry(), MegatronMappingRegistry)
+
+    def test_has_mappings(self, bridge):
+        assert len(bridge.mapping_registry().mappings) > 0
 
     def test_has_embeddings_mapping(self, bridge):
         names = self._collect_names(bridge.mapping_registry())
@@ -311,6 +738,15 @@ class TestGemma4VLBridgeMappingRegistry:
         names = self._collect_names(bridge.mapping_registry())
         assert any("embed_vision" in n for n in names)
 
+    def test_has_audio_tower_mapping(self, bridge):
+        """VL bridge includes audio_tower mappings."""
+        names = self._collect_names(bridge.mapping_registry())
+        assert any("audio_tower" in n for n in names)
+
+    def test_has_embed_audio_mapping(self, bridge):
+        names = self._collect_names(bridge.mapping_registry())
+        assert any("embed_audio" in n for n in names)
+
     def test_has_qkv_mapping(self, bridge):
         names = self._collect_names(bridge.mapping_registry())
         assert any("linear_qkv" in n for n in names)
@@ -320,7 +756,6 @@ class TestGemma4VLBridgeMappingRegistry:
         assert any("mlp" in n for n in names)
 
     def test_has_shared_expert_layernorm(self, bridge, mock_hf_config_moe):
-        # MoE-specific mappings require hf_config to be set
         bridge.hf_config = mock_hf_config_moe
         names = self._collect_names(bridge.mapping_registry())
         assert any("post_shared_expert_layernorm" in n for n in names)
@@ -330,38 +765,36 @@ class TestGemma4VLBridgeMappingRegistry:
         names = self._collect_names(bridge.mapping_registry())
         assert any("post_moe_layernorm" in n for n in names)
 
-
-# ---------------------------------------------------------------------------
-# Edge cases
-# ---------------------------------------------------------------------------
+    def test_uses_language_model_prefix_for_vl(self, bridge, mock_hf_config_moe):
+        """VLM uses model.language_model.layers.* (not model.layers.*)."""
+        bridge.hf_config = mock_hf_config_moe
+        names = self._collect_names(bridge.mapping_registry())
+        lm_keys = [n for n in names if "layers" in n and "vision" not in n and "audio" not in n]
+        assert any("language_model" in n for n in lm_keys)
 
 
 class TestGemma4VLBridgeEdgeCases:
     def test_custom_token_ids(self, bridge, mock_hf_pretrained_moe):
         mock_hf_pretrained_moe.config.image_token_id = 99999
         mock_hf_pretrained_moe.config.bos_token_id = 42
-        provider = bridge.provider_bridge(mock_hf_pretrained_moe)
-        assert provider.image_token_id == 99999
-        assert provider.bos_token_id == 42
+        p = bridge.provider_bridge(mock_hf_pretrained_moe)
+        assert p.image_token_id == 99999
+        assert p.bos_token_id == 42
 
     def test_default_image_token_id(self, bridge, mock_hf_pretrained_moe):
         del mock_hf_pretrained_moe.config.image_token_id
-        provider = bridge.provider_bridge(mock_hf_pretrained_moe)
-        assert provider.image_token_id == 258_880
+        assert bridge.provider_bridge(mock_hf_pretrained_moe).image_token_id == 258_880
 
     def test_default_vision_soft_tokens(self, bridge, mock_hf_pretrained_moe):
         del mock_hf_pretrained_moe.config.vision_soft_tokens_per_image
-        provider = bridge.provider_bridge(mock_hf_pretrained_moe)
-        assert provider.vision_soft_tokens_per_image == 280
+        assert bridge.provider_bridge(mock_hf_pretrained_moe).vision_soft_tokens_per_image == 280
 
     def test_different_vocab_sizes(self, bridge, mock_hf_pretrained_moe):
-        for vocab_size in [256000, 262144, 300000]:
-            mock_hf_pretrained_moe.config.text_config.vocab_size = vocab_size
-            provider = bridge.provider_bridge(mock_hf_pretrained_moe)
-            assert provider.vocab_size == vocab_size
+        for vs in [256000, 262144, 300000]:
+            mock_hf_pretrained_moe.config.text_config.vocab_size = vs
+            assert bridge.provider_bridge(mock_hf_pretrained_moe).vocab_size == vs
 
     def test_different_layer_counts(self, bridge, mock_hf_pretrained_moe):
-        for num_layers in [32, 46, 62]:
-            mock_hf_pretrained_moe.config.text_config.num_hidden_layers = num_layers
-            provider = bridge.provider_bridge(mock_hf_pretrained_moe)
-            assert provider.num_layers == num_layers
+        for nl in [32, 46, 62]:
+            mock_hf_pretrained_moe.config.text_config.num_hidden_layers = nl
+            assert bridge.provider_bridge(mock_hf_pretrained_moe).num_layers == nl
